% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3gbi-package.R
\docType{package}
\name{b3gbi-package}
\alias{b3gbi}
\alias{b3gbi-package}
\title{b3gbi: General Biodiversity Indicators for Biodiversity Data Cubes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate general biodiversity indicators from species occurrence cubes, such as those produced by the 'GBIF' (Global Biodiversity Information Facility) occurrence cube service \url{https://techdocs.gbif.org/en/data-use/data-cubes}. Indicators include observed and coverage-based rarefied species richness and Hill numbers (via 'iNEXT'), evenness, rarity, taxonomic distinctness, occupancy, turnover and sample completeness, and can be calculated over time (time series) or space (maps), with bootstrap confidence intervals. Cubes using the EEA reference grid, MGRS, the extended quarter-degree grid (EQDGC) or ISEA3H grids are supported. For details see Dove (2026) \doi{10.21105/joss.10963}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-cubed-eu/b3gbi}
  \item \url{https://b-cubed-eu.github.io/b3gbi/}
  \item Report bugs at \url{https://github.com/b-cubed-eu/b3gbi/issues}
}

}
\author{
\strong{Maintainer}: Shawn Dove \email{Shawn.Dove@allzool.bio.uni-giessen.de} (\href{https://orcid.org/0000-0001-9465-5638}{ORCID})

Other contributors:
\itemize{
  \item Ward Langeraert (\href{https://orcid.org/0000-0002-5900-8109}{ORCID}) [contributor]
  \item Peter Desmet (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) [contributor]
  \item University of Giessen [copyright holder]
  \item European Union [funder]
}

}
\keyword{internal}
