% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci_methods.R
\name{calc_ci.evenness_core}
\alias{calc_ci.evenness_core}
\title{Core function to calculate confidence intervals for evenness. This is called
by the calc_ci.pielou_evenness and calc_ci.williams_evenness functions.}
\usage{
\method{calc_ci}{evenness_core}(x, type, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)
}
\arguments{
\item{x}{A data cube object}

\item{type}{Evenness measure ("pielou_evenness" or "williams_evenness)}

\item{indicator}{An indicator calculated over time, in the form of a data
frame. *Note: this should NOT be an 'indicator_ts' object as it is meant to
be called by the 'compute_indicator_workflow' function.}

\item{num_bootstrap}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}

\item{ci_type}{Type of bootstrap confidence intervals to calculate.
(Default: "norm". Select "none" to avoid calculating bootstrap CIs.)}
}
\description{
Core function to calculate confidence intervals for evenness. This is called
by the calc_ci.pielou_evenness and calc_ci.williams_evenness functions.
}
