% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci_methods.R
\name{calc_ci.hill_core}
\alias{calc_ci.hill_core}
\title{Core function for handling the confidence interval calculations for Hill
diversity. This function is called by the calc_ci.hill0, calc_ci.hill1,
calc_ci.hill2 functions.}
\usage{
\method{calc_ci}{hill_core}(x, indicator, type = c("hill0", "hill1", "hill2"), ...)
}
\arguments{
\item{x}{A data cube object}

\item{indicator}{An indicator calculated over time, in the form of a data
frame. *Note: this should NOT be an 'indicator_ts' object as it is meant to
be called by the 'compute_indicator_workflow' function.}

\item{type}{Type of Hill diversity function to calculate confidence
intervals for. Options are "hill0" for richness, "hill1" for Simpson-Hill
diversity, or "hill2" for Shannon-Hill diversity.}

\item{...}{Additional arguments passed to specific indicator calculation
functions.}
}
\description{
Core function for handling the confidence interval calculations for Hill
diversity. This function is called by the calc_ci.hill0, calc_ci.hill1,
calc_ci.hill2 functions.
}
