% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calc_indicator}
\alias{calc_indicator}
\title{Calculate a Biodiversity Indicator}
\usage{
calc_indicator(x)
}
\arguments{
\item{x}{An object containing the data required for the indicator calculation.}

\item{...}{Additional arguments (potentially used by specific methods).}
}
\value{
The calculated biodiversity indicator value(s). The format of the
output will depend on the specific S3 method used.
}
\description{
A generic function for calculating biodiversity indicators.
Specific implementations for different indicator types should be provided
using S3 methods.
}
