% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rarity.R
\name{calc_rarity}
\alias{calc_rarity}
\title{Calculate rarity trend}
\usage{
calc_rarity(
  data,
  method = "grid",
  type = "area",
  cs1 = 100,
  cs2 = 100,
  level = "country",
  region = "Denmark"
)
}
\arguments{
\item{data}{A tibble created from a GBIF cube using the process_cube function.}

\item{method}{A character vector.
"observed" calculates the observed rarity trend from the cube data,}
}
\value{
A tibble
}
\description{
Calculate rarity trend
}
