% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicator.R
\name{calculate_indicator}
\alias{calculate_indicator}
\title{Calculate Gridded Biodiversity Indicators}
\usage{
calculate_indicator(
  x,
  type,
  cs1 = 100,
  cs2 = 100,
  level = c("continent", "country", "world"),
  region = "Europe",
  ...
)
}
\arguments{
\item{x}{A data cube object ('processed_cube', 'processed_cube_dsinfo', or  'virtual_cube').}

\item{type}{The indicator to calculate. Supported options include:
* 'hill0', 'hill1', 'hill2': Hill numbers (order 0, 1, and 2).
* 'obs_richness': Observed species richness.
* 'total_occ': Total number of occurrences.
* 'newness': Mean year of occurrence.
* 'density': Density of occurrences.
* 'e9_evenness', 'pielou_evenness': Evenness measures.
* 'ab_rarity', 'area_rarity':  Abundance-based and area-based rarity scores.
* 'spec_occ': Species occurrences.
* 'tax_distinct': Taxonomic distinctness.}

\item{cs1}{Length of the grid cell's sides, in kilometers, if square. (Default: 100)}

\item{cs2}{Width of the grid cell's sides, in kilometers. (Default: 100)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{...}{Additional arguments passed to specific indicator calculation functions.}
}
\value{
An S3 object of the appropriate class containing the calculated indicator values and metadata:
  * 'indicator_map' for real-world observational data.
  * 'virtual_indicator_map' for virtual species data.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators on a spatial grid. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_map' or 'virtual_indicator_map').
}
\examples{
# Assuming 'my_data_cube' is a 'processed_cube' or 'virtual_cube' object
diversity_map <- calculate_indicator(my_data_cube, type = "obs_richness", level = "continent", region = "Africa")

}
