% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_crs_resolution.R
\name{create_grid_from_reprojected_data}
\alias{create_grid_from_reprojected_data}
\title{Create a grid based on reprojected data}
\usage{
create_grid_from_reprojected_data(data, resolution)
}
\arguments{
\item{data}{Reprojected spatial data (sf or terra object)}

\item{resolution}{Cell size in the units of data's CRS c(x, y)}
}
\value{
An sf object with the grid
}
\description{
Create a grid based on reprojected data
}
