% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_crs_resolution.R
\name{determine_cell_size}
\alias{determine_cell_size}
\title{Determine appropriate cell size when changing CRS}
\usage{
determine_cell_size(
  data,
  input_resolution,
  input_crs,
  target_crs,
  target_resolution = NULL,
  input_units,
  target_units
)
}
\arguments{
\item{data}{Input spatial data (sf or terra object)}

\item{input_resolution}{Original resolution as vector c(x, y)}

\item{input_crs}{CRS of the input data (EPSG code or proj4string)}

\item{target_crs}{Target CRS (EPSG code or proj4string)}

\item{target_resolution}{Optional. If provided, validates if it's appropriate}
}
\value{
A list with recommended cell size in target CRS units and validation result
}
\description{
Determine appropriate cell size when changing CRS
}
