% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cube_2}
\alias{example_cube_2}
\title{Cube of GBIF Insect Occurrences in Europe}
\format{
A 'processed_cube' object containing a tibble with 23,025 rows and 10 variables, as well as metadata
\describe{
\item{year}{year occurrence was recorded}
\item{eea_cell_code}{code containing the cell resolution and coordinates on the EEA grid}
\item{taxonKey}{taxonomic key associated with the species on GBIF}
\item{obs}{number of individuals observed}
\item{scientificName}{scientific species name}
\item{rank}{taxonomic rank}
\item{kingdom}{kingdom}
\item{xcoord}{East-West coordinate on the EEA grid}
\item{ycoord}{North-South coordinate on the EEA grid}
\item{resolution}{grid cell size}
}
}
\source{
\url{https://doi.org/10.15468/dl.qk4f2z}
}
\usage{
example_cube_2
}
\description{
Small example cube containing the insect occurrences in Europe available on GBIF as of 16.03.2024.
}
\examples{
{
    \dontrun{
      # Basic example of how to use the dataset
      europe_insects_or_map <- obs_richness_map(example_cube_2, level = "continent", region = "Europe")
      plot(europe_insects_or_map, title = "Insects in Europe (1751-2023): Observed Species Richness")
    }
  }
}
\keyword{datasets}
