% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_indicator_map2}
\alias{example_indicator_map2}
\title{Map of Abundance-Based Rarity for Insects in Europe}
\format{
An 'indicator_map' object containing a tibble with 3,225 rows and 4 variables, as well as metadata
\describe{
\item{cellid}{ID number identifying a particular grid cell}
\item{area_km2}{area of the grid cell in square kms}
\item{diversity_val}{calculated rarity value for the grid cell}
\item{geometry}{grid cell polygon (this column is not visible by default when you print the object)}
}
}
\source{
\url{https://doi.org/10.15468/dl.qk4f2z}
}
\usage{
example_indicator_map2
}
\description{
Example indicator containing a map of abundance-based rarity for insect occurrences in Europe (occurrences from GBIF: 16.03.2024).
}
\keyword{datasets}
