% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs_for_mgrs.R
\name{get_crs_for_mgrs}
\alias{get_crs_for_mgrs}
\title{Get the CRS for MGRS cell codes}
\usage{
get_crs_for_mgrs(cellcodes)
}
\arguments{
\item{cellcodes}{A character vector of MGRS grid cell codes (e.g., "32UMA",
"33TWE").}
}
\value{
A character string representing the most probable EPSG code for the
given cell codes, for example, \code{"EPSG:32632"}.
}
\description{
This function takes a vector of MGRS (Military Grid Reference System) cell
codes and determines the most likely UTM (Universal Transverse Mercator)
coordinate reference system (CRS) in EPSG format. It extracts the UTM zone
and determines the hemisphere (North or South) based on the letter codes.
}
\examples{
# Example with MGRS codes from a Northern Hemisphere zone (Zone 32)
mgrs_codes_north <- c("32TMT", "32TMR", "32TML")
get_crs_for_mgrs(mgrs_codes_north)

# Example with MGRS codes from a Southern Hemisphere zone (Zone 14)
mgrs_codes_south <- c("14KMF", "14KMJ", "14KMG")
get_crs_for_mgrs(mgrs_codes_south)

# The function can handle a mix of zones, choosing the median
mgrs_mixed_zones <- c("27TMT", "28TMR", "29TWE")
get_crs_for_mgrs(mgrs_mixed_zones)

}
