% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed_years.R
\name{get_observed_years}
\alias{get_observed_years}
\title{Extract Years With Observations from an Indicator Map}
\usage{
get_observed_years(x)
}
\arguments{
\item{x}{An "indicator_map" object containing calculated indicator values
associated with grid cells.}
}
\value{
A data frame with two columns:
\itemize{
\item \code{years}: A sequence of years covering the range of observations.
\item \code{occurrences}: A logical vector indicating if observations exist
for each year (\code{TRUE} if present, \code{FALSE} if absent).
}
}
\description{
Takes an "indicator_map",  "indicator_ts", or "processed_cube" object and
determines the years for which observation data exists.
}
\examples{
\donttest{
total_occ_mammals_denmark <- total_occ_map(example_cube_1, level = "country",
                                           region = "Denmark")
get_observed_years(total_occ_mammals_denmark)
}
}
