% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_species.R
\name{list_species}
\alias{list_species}
\title{Extract Species Names.}
\usage{
list_species(object)
}
\arguments{
\item{object}{A biodiversity data object containing species names (either in
a \code{data} slot or a \code{species_names} vector).}
}
\value{
A tibble with a single column:
\itemize{
\item \code{scientificName}: The unique scientific names found in the object.
}
}
\description{
Retrieves a standardized list of species names from a biodiversity data object
('processed_cube', 'indicator_map' or 'indicator_ts').
}
\examples{
list_species(example_cube_1)
}
