% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_crs_resolution.R
\name{meters_per_degree}
\alias{meters_per_degree}
\title{Calculate approximate meters per degree at a given latitude}
\usage{
meters_per_degree(latitude)
}
\arguments{
\item{latitude}{Latitude in decimal degrees}
}
\value{
Named vector with meters per degree longitude and latitude
}
\description{
Calculate approximate meters per degree at a given latitude
}
