% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_virtual_cube}
\alias{new_virtual_cube}
\title{'virtual_cube' S3 Constructor}
\usage{
new_virtual_cube(x)
}
\arguments{
\item{x}{A data.table containing virtual species occurrence data. Required columns include:
\itemize{
\item 'year'
\item 'taxonKey'
\item 'obs' (number of occurrences)
\item 'scientificName'
\item 'xcoord', 'ycoord' (spatial coordinates)
}}
}
\value{
A 'virtual_cube' S3 object containing:
\itemize{
\item \strong{Summary statistics:} First and last year (if applicable), coordinate range,
number of spatial cells, species, and observations.
\item \strong{Original data:} The input data.table.
}
}
\description{
This function creates a 'virtual_cube' S3 object, a specialized structure
for holding and analyzing virtual species occurrence data.  Since this data might not
include all the biodiversity details of field data, it focuses on the core components
for spatial analysis.
}
\note{
'virtual_cube' objects are used by various analysis functions within this package.
}
\examples{
# Assuming your virtual species data is in 'simulated_data'
my_virtual_cube <- new_virtual_cube(simulated_data)

}
