% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_virtual_indicator_map}
\alias{new_virtual_indicator_map}
\title{'virtual_indicator_map' S3 Constructor}
\usage{
new_virtual_indicator_map(
  x,
  div_type,
  cs1,
  cs2,
  map_level,
  map_region,
  num_species,
  first_year,
  last_year,
  num_years
)
}
\arguments{
\item{x}{An sf data frame containing at least two columns:
\itemize{
\item 'cellid': Unique ID for each spatial cell.
\item 'diversity_val': The calculated indicator value for the cell.
}}

\item{div_type}{The type of biodiversity indicator (e.g., "obs_richness").}

\item{cs1}{Length of the cell's sides, in kilometers, if square.}

\item{cs2}{Width of the cells sides, in kilometers. Use only if the cell is non-square.}

\item{map_level}{The spatial level of the map (e.g., "country", "continent", "world").}

\item{map_region}{The name of the spatial region under analysis.}

\item{num_species}{The total number of virtual species included in the analysis.}

\item{first_year}{The first year of the indicator calculation timeframe.}

\item{last_year}{The last year of the indicator calculation timeframe.}

\item{num_years}{The number of years in the time series.}
}
\value{
A 'virtual_indicator_map' S3 object containing:
\itemize{
\item \strong{Indicator name and type:}  Descriptive name, calculated type.
\item \strong{Cell information:} Number of cells, cell size (area).
\item \strong{Spatial details:} Level (cell, country), region name, projection, and coordinate range.
\item \strong{Virtual data info:} Number of species.
\item \strong{Years analyzed:} First and last year, number of years.
\item \strong{Mapped Results:} The input sf object, now containing indicator scores.
}
}
\description{
This function creates a 'virtual_indicator_map' S3 object, a specialized
structure for storing spatial biodiversity indicator results calculated on virtual species
data. It stores metadata and calculated values mapped onto geographic cells.
}
\examples{
# Assuming results exist in 'vsim_indicator' and other parameters are defined
my_virtual_map <- new_virtual_indicator_map(vsim_indicator, "obs_richness", ...)

}
