% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{newness_ts}
\alias{newness_ts}
\title{Calculate Mean Year of Occurrence Over Time}
\usage{
newness_ts(data, ...)
}
\arguments{
\item{x}{A data cube object (class 'processed_cube').}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_ts' and 'newness' containing
the calculated indicator values and metadata.
}
\description{
This function estimates the change in relative newness of records
in a data cube over time by calculating the mean year of occurrence as a time series.
It prepares the data, creates a grid, calculates the indicator value for each
grid cell, and formats the output into an S3 object.
}
\examples{
n_ts <- newness_ts(example_cube_1)
plot(n_ts)

}
