% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{occ_density_ts}
\alias{occ_density_ts}
\title{Calculate Occurrence Density Over Time}
\usage{
occ_density_ts(data, ...)
}
\arguments{
\item{x}{A data cube object (class 'processed_cube').}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_ts' and 'occ_density' containing
the calculated indicator values and metadata.
}
\description{
This function calculates density of records as a time series.
}
\examples{
od_ts <- occ_density_ts(example_cube_1, first_year = 1985)
plot(od_ts)

}
