% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{pielou_evenness_ts}
\alias{pielou_evenness_ts}
\title{Calculate Pielou's Evenness Over Time}
\usage{
pielou_evenness_ts(data, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{cube_crs}}{The projection of the cube. (Default: "EPSG:3035")}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
  }}
}
\value{
An S3 object with the classes 'indicator_ts' and 'pielou_evenness' containing
the calculated indicator values and metadata.
}
\description{
This function calculates Pielou's evenness over time.
}
\examples{
pe_ts <- pielou_evenness_ts(example_cube_1, first_year = 1985)
plot(pe_ts)

}
\seealso{
compute_indicator_workflow
}
