% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.ab_rarity}
\alias{plot.ab_rarity}
\title{Plot Abundance-Based Rarity}
\usage{
\method{plot}{ab_rarity}(x, ...)
}
\arguments{
\item{x}{An object containing abundance-based rarity data. Must be of
class 'ab_rarity', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots abundance-based rarity data, either as a time series
('indicator_ts' object)  or as a spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'ab_rarity_ts' (indicator_ts) and 'ab_rarity_map' (indicator_map)
# Time series plot:
plot.ab_rarity(ab_rarity_ts)

# Map visualization:
plot.ab_rarity(ab_rarity_map)

}
