% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.cum_richness}
\alias{plot.cum_richness}
\title{Plot Cumulative Species Richness}
\usage{
\method{plot}{cum_richness}(x, auccolour = NULL, ...)
}
\arguments{
\item{x}{An object containing cumulative species richness data. Must be of
class 'cum_richness' and an 'indicator_ts' object.}

\item{auccolour}{(Optional). Colour for the area under the curve of a
cumulative richness plot. Default is orange.}

\item{...}{Additional arguments passed to the internal plotting function
(\code{plot_ts}). See its documentation for details.}
}
\value{
A ggplot object representing a time series plot of cumulative
species richness.
}
\description{
Plots cumulative species richness data as a time series
using an 'indicator_ts' object.
}
\examples{
# Assuming you have an 'indicator_ts' object named 'cum_richness_ts'
plot.cum_richness(cum_richness_ts)

}
