% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.hill0}
\alias{plot.hill0}
\title{Plot Species Richness Index (Hill Number 0)}
\usage{
\method{plot}{hill0}(x, ...)
}
\arguments{
\item{x}{An object containing species richness index (Hill number 0) data.
Must be of class 'hill0', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots species richness index data (estimated by coverage-based
rarefaction), either as a time series ('indicator_ts' object)  or as a
spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'hill0_ts' (indicator_ts) and 'hill0_map' (indicator_map)
# Time series plot:
plot.hill0(hill0_ts)

# Map visualization:
plot.hill0(hill0_map)

}
