% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.hill1}
\alias{plot.hill1}
\title{Plot Hill-Shannon Diversity Index (Hill Number 1)}
\usage{
\method{plot}{hill1}(x, ...)
}
\arguments{
\item{x}{An object containing Hill-Shannon diversity index (Hill number 1) data.
Must be of class 'hill1', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots Hill-Shannon diversity index data (estimated by coverage-based
rarefaction), either as a time series ('indicator_ts' object)  or as a
spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'hill1_ts' (indicator_ts) and 'hill1_map' (indicator_map)
# Time series plot:
plot.hill1(hill1_ts)

# Map visualization:
plot.hill1(hill1_map)

}
