% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.hill2}
\alias{plot.hill2}
\title{Plot Hill-Simpson Diversity Index (Hill Number 2)}
\usage{
\method{plot}{hill2}(x, ...)
}
\arguments{
\item{x}{An object containing Hill-Simpson diversity index (Hill number 2) data.
Must be of class 'hill2', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots Hill-Simpson diversity index data (estimated by coverage-based
rarefaction), either as a time series ('indicator_ts' object)  or as a
spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'hill2_ts' (indicator_ts) and 'hill2_map' (indicator_map)
# Time series plot:
plot.hill2(hill2_ts)

# Map visualization:
plot.hill2(hill2_map)

}
