% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.newness}
\alias{plot.newness}
\title{Plot Mean Year of Occurrence}
\usage{
\method{plot}{newness}(x, ...)
}
\arguments{
\item{x}{An object containing mean year of occurrence associated with map
grid cells. Must be of class 'newness' and an 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting function
(\code{plot_map}). See its documentation for details.}
}
\value{
A ggplot object representing a map of the mean year of occurrences.
}
\description{
Creates a spatial map visualization of the mean year of
species occurrences.
}
\examples{
# Assuming you have an 'indicator_map' object named 'newness_map'
plot.newness(newness_map)

}
