% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.obs_richness}
\alias{plot.obs_richness}
\title{Plot Observed Species Richness}
\usage{
\method{plot}{obs_richness}(x, ...)
}
\arguments{
\item{x}{An object containing observed species richness data. Must be of
class 'obs_richness', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots observed species richness data, either as a time series
('indicator_ts' object)  or as a spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'richness_ts' (indicator_ts) and 'richness_map' (indicator_map)
# Time series plot:
plot.obs_richness(richness_ts)

# Map visualization:
plot.obs_richness(richness_map)

}
