% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.occ_by_dataset}
\alias{plot.occ_by_dataset}
\title{Plot Occurrences Segregated by Dataset}
\usage{
\method{plot}{occ_by_dataset}(
  x,
  x_breaks = 6,
  facet_scales = "free_y",
  facet_rows = NULL,
  facet_cols = NULL,
  facet_label_width = 60,
  max_datasets = 20,
  min_occurrences = NULL,
  ...
)
}
\arguments{
\item{x}{An object containing occurrence data segregated by dataset. Must
be of class 'occ_by_dataset' and an 'indicator_ts' object.}

\item{x_breaks}{(Optional) Integer giving desired number of breaks for the x-axis.
(May not return exactly the number requested.)}

\item{facet_scales}{Controls y-axis scaling across facets. Use "free_y" to
allow independent scaling, or "fixed" for a common scale.}

\item{facet_rows}{(Optional) Number of rows of facets.}

\item{facet_cols}{(Optional) Number of columns of facets.}

\item{facet_label_width}{(Optional) Controls the maximum width of facet labels.}

\item{max_datasets}{(Optional) Maximum number of datasets to include in the plot.
Datasets are selected based on having the highest number of occurrences.
*Note that increasing this too much could result in high memory usage and/or
make the plot very difficult to read.}

\item{min_occurrences}{(Optional)  Minimum total number of occurrences for a
dataset to be included in the plot.}

\item{...}{Additional arguments passed to the internal plotting function
(\code{plot_ts_seg}). See its documentation for details.}
}
\value{
A ggplot object representing a time series plot with occurrences
segregated by dataset.
}
\description{
Creates a time series plot of total occurrences, with the
occurrences visually segregated by their dataset of origin. Requires an
'indicator_ts' object.
}
\examples{
# Assuming you have an 'indicator_ts' object named 'occ_by_dataset_ts'
plot.occ_by_dataset(occ_by_dataset_ts)

# Only plot datasets with at least 100 occurrences:
plot.occ_by_dataset(occ_by_dataset_ts, min_occurrences = 100)

}
