% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.occ_by_type}
\alias{plot.occ_by_type}
\title{Plot Occurrences Segregated by Type}
\usage{
\method{plot}{occ_by_type}(
  x,
  x_breaks = 6,
  facet_scales = "free_y",
  facet_rows = NULL,
  facet_cols = NULL,
  facet_label_width = 60,
  ...
)
}
\arguments{
\item{x}{An object containing  occurrence data segregated by type. Must
be of class 'occ_by_type' and an 'indicator_ts' object.}

\item{x_breaks}{(Optional)  Integer giving desired number of breaks for the x-axis.
(May not return exactly the number requested.)}

\item{facet_scales}{Controls y-axis scaling across facets.  Use "free_y" to
allow independent scaling, or "fixed" for a common scale.}

\item{facet_rows}{(Optional) Number of rows of facets.}

\item{facet_cols}{(Optional) Number of columns of facets.}

\item{facet_label_width}{(Optional) Controls the maximum width of facet labels.}

\item{...}{Additional arguments passed to the internal plotting function
(\code{plot_ts_seg}). See its documentation for details.}
}
\value{
A ggplot object representing a time series plot with occurrences
segregated by type.
}
\description{
Creates a time series plot of total occurrences, with the
occurrences visually segregated by their type. Requires an 'indicator_ts'
object.
}
\examples{
# Assuming you have an 'indicator_ts' object named 'occ_by_type_ts'
plot.occ_by_type(occ_by_type_ts)

}
