% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.occ_density}
\alias{plot.occ_density}
\title{Plot Occurrence Density}
\usage{
\method{plot}{occ_density}(x, ...)
}
\arguments{
\item{x}{An object containing occurrence density data. Must be of
class 'density', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots occurrence density data, either as a time series
('indicator_ts' object)  or as a spatial map ('indicator_map' object).
}
\examples{
# Assuming objects 'density_ts' (indicator_ts) and 'density_map' (indicator_map)
# Time series plot:
plot.density(density_ts)

# Map visualization:
plot.density(density_map)

}
