% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.rarefied}
\alias{plot.rarefied}
\title{Plot Rarefied Species Richness}
\usage{
\method{plot}{rarefied}(x, ...)
}
\arguments{
\item{x}{An object containing rarefied species richness data. Must be of
class 'rarefied', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots rarefied species richness data (estimated by sample
size-based rarefaction), either as a time series ('indicator_ts' object)
or as a spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'rarefied_ts' (indicator_ts) and 'rarefied_map' (indicator_map)
# Time series plot:
plot.rarefied(rarefied_ts)

# Map visualization:
plot.rarefied(rarefied_map)

}
