% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.tax_distinct}
\alias{plot.tax_distinct}
\title{Plot Taxonomic Distinctness}
\usage{
\method{plot}{tax_distinct}(x, ...)
}
\arguments{
\item{x}{An object containing taxonomic distinctness data. Must be of
class 'tax_distinct', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots taxonomic distinctness data, either as a time series
('indicator_ts' object)  or as a spatial map ('indicator_map' object).
}
\examples{
# Assuming objects 'tax_distinct_ts' (indicator_ts) and 'tax_distinct_map' (indicator_map)
# Time series plot:
plot.tax_distinct(tax_distinct_ts)

# Map visualization:
plot.tax_distinct(tax_distinct_map)

}
