% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.total_occ}
\alias{plot.total_occ}
\title{Plot Total Occurrences}
\usage{
\method{plot}{total_occ}(x, ...)
}
\arguments{
\item{x}{An object containing summed occurrence data. Must be of
class 'total_occ', either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots summed occurrence data, either as a time series
('indicator_ts' object)  or as a spatial map ('indicator_map' object).
}
\examples{
# Assuming you have objects 'total_occ_ts' (indicator_ts) and 'total_occ_map' (indicator_map)
# Time series plot:
plot.total_occ(total_occ_ts)

# Map visualization:
plot.total_occ(total_occ_map)

}
