% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.williams_evenness}
\alias{plot.williams_evenness}
\title{Plot Williams' Evenness}
\usage{
\method{plot}{williams_evenness}(x, ...)
}
\arguments{
\item{x}{An object containing evenness data. Must be of class 'evenness',
either an 'indicator_ts' or 'indicator_map' object.}

\item{...}{Additional arguments passed to the internal plotting functions
(\code{plot_ts} or \code{plot_map}). See their documentation for details.}
}
\value{
A ggplot object representing either a time series plot or a map,
depending on the class of the input object \code{x}.
}
\description{
Plots evenness data,
either as a time series ('indicator_ts' object)  or as a spatial map
('indicator_map' object).
}
\examples{
# Assuming objects 'evenness_ts' (indicator_ts) and 'evenness_map' (indicator_map)
# Time series plot:
plot.evenness(evenness_ts)

# Map visualization:
plot.evenness(evenness_map)

}
