% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.indicator_ts}
\alias{print.indicator_ts}
\title{Print an Indicator Time Series Object}
\usage{
\method{print}{indicator_ts}(x, n = 10, ...)
}
\arguments{
\item{x}{An indicator_ts object.}

\item{n}{(Optional) Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a summary representation of an indicator_ts object,
designed for user-friendly display in the console.
}
\examples{
print(example_indicator_ts1)

}
