% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.processed_cube_dsinfo}
\alias{print.processed_cube_dsinfo}
\title{Print a Processed Data Cube Object with Dataset Information}
\usage{
\method{print}{processed_cube_dsinfo}(x, n = 10, ...)
}
\arguments{
\item{x}{A processed_cube_dsinfo object.}

\item{n}{(Optional) Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments passed to print().}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a summary representation of a processed_cube_dsinfo
object, designed for user-friendly display in the console.
}
