% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.sim_cube}
\alias{print.sim_cube}
\title{Print a Simulated Data Cube Object}
\usage{
\method{print}{sim_cube}(x, n = 10, ...)
}
\arguments{
\item{x}{A sim_cube object.}

\item{n}{(Optional) Integer specifying the number of rows of cube data to
display.}

\item{...}{Additional arguments passed to print().}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a summary representation of a sim_cube object,
designed for user-friendly display in the console.
}
