% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.virtual_cube}
\alias{print.virtual_cube}
\title{Print a Virtual Cube Object}
\usage{
\method{print}{virtual_cube}(x, n = 5, ...)
}
\arguments{
\item{x}{A virtual_cube object.}

\item{n}{Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments.}
}
\description{
Provides a summary representation of a virtual_cube object,
designed for user-friendly display in the console.
}
