% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{relative_occupancy_ts}
\alias{relative_occupancy_ts}
\title{Calculate Species Relative Occupancy Over Time}
\usage{
relative_occupancy_ts(data, ...)
}
\arguments{
\item{data}{A processed data cube (processed_cube object).}

\item{...}{Additional arguments passed to \code{\link{compute_indicator_workflow}}.}
}
\value{
An 'indicator_ts' object with class 'relative_occupancy' containing:
\itemize{
\item \strong{year}: Years.
\item \strong{diversity_val}: Relative occupancy values (0-1).
}
}
\description{
This function calculates the relative occupancy of each species over time
by dividing the number of grid cells occupied by the species in each year
by the total number of cells in the study area (including cells with no data).
}
\seealso{
Other species-based indicators: 
\code{\link{relative_occupancy_map}()}
}
\concept{species-based indicators}
