% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_crs_resolution.R
\name{reproject_and_create_grid}
\alias{reproject_and_create_grid}
\title{Example workflow function to demonstrate the entire process}
\usage{
reproject_and_create_grid(
  input_data,
  input_resolution,
  target_crs,
  target_resolution = NULL,
  input_units,
  target_units
)
}
\arguments{
\item{input_data}{Input spatial data (sf or terra object)}

\item{input_resolution}{Resolution of input data c(x, y)}

\item{target_crs}{Target CRS as EPSG code or proj4string}

\item{target_resolution}{Optional target resolution in target CRS units}
}
\value{
A list with the reprojected data and grid
}
\description{
Example workflow function to demonstrate the entire process
}
