% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mv.R
\name{save_mv}
\alias{save_mv}
\title{Save mapview maps}
\usage{
save_mv(x, file, ...)
}
\arguments{
\item{x}{An 'indicator_map' object containing multi-species indicator values
associated with map grid cells. This is a required parameter with no
default.}

\item{file}{File name and path to store the map. The extension will
determine the file type.}

\item{...}{Additional arguments passed to the mapshot2 function.}
}
\description{
Save mapview maps as interactive .html files, or snapshots as
.png, pdf, or .jpeg files.
}
