% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_occurrences.R
\name{species_occurrences}
\alias{species_occurrences}
\title{Calculate occurrences for individual species over time or across a grid}
\usage{
species_occurrences(
  data,
  method = "grid",
  level = "continent",
  region = "Europe",
  cs1 = 100,
  cs2 = 100
)
}
\arguments{
\item{data}{A tibble created from a GBIF cube using the process_cube function.}

\item{method}{A character vector.}
}
\value{
?
}
\description{
Calculate occurrences for individual species over time or across a grid
}
