% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_range.R
\name{species_range}
\alias{species_range}
\title{Calculate occurrences for individual species over time or across a grid}
\usage{
species_range(data, level = "country", region = "Germany", cs1 = 10, cs2 = 10)
}
\arguments{
\item{data}{A tibble created from a GBIF cube using the process_cube function.}

\item{method}{A character vector.}
}
\value{
?
}
\description{
Calculate occurrences for individual species over time or across a grid
}
