% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{tax_distinct_ts}
\alias{tax_distinct_ts}
\title{Calculate Taxonomic Distinctness Over Time}
\usage{
tax_distinct_ts(data, rows = 1, ...)
}
\arguments{
\item{rows}{Choose which row to select if there are multiple matches when retrieving taxonomic information from GBIF. (Default is 1. Use NA for interactive mode.)}

\item{x}{A data cube object (class 'processed_cube').}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_ts' and 'tax_distinct' containing
the calculated indicator values and metadata.
}
\description{
This function calculates the taxonomic distinctness index as a time series.
}
\examples{
td_ts <- tax_distinct_ts(example_cube_1, level = "country", region = "Denmark")
plot(td_ts)

}
