% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{total_occ_map}
\alias{total_occ_map}
\alias{total_occ_ts}
\title{Calculate Total Occurrences Over Space or Time}
\usage{
total_occ_map(data, ...)

total_occ_ts(data, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{cell_size}}{(Optional) Length of grid cell sides, in km or degrees.
If set to "grid" (default), this will use the existing grid size of your
cube. If set to "auto", this will be automatically determined according to
the geographical level selected. This is 100 km or 1 degree for 'continent'
or 'world', 10 km or (for a degree-based CRS) the native resolution of the
cube for 'country', 'sovereignty' or 'geounit'. If level is set to 'cube',
cell size will be the native resolution of the cube for a degree-based CRS,
or for a km-based CRS, the cell size will be determined by the area of the
cube: 100 km for cubes larger than 1 million sq km, 10 km for cubes between
10 thousand and 1 million sq km, 1 km for cubes between 100 and 10 thousand
sq km, and 0.1 km for cubes smaller than 100 sq km. Alternatively, the user
can manually select the grid cell size (in km or degrees). Note that the
cell size must be a whole number multiple of the cube's resolution.}
    \item{\code{level}}{(Optional) Spatial level: 'cube', 'continent', 'country',
'world', 'sovereignty', or 'geounit'. (Default: 'cube')}
    \item{\code{region}}{(Optional) The region of interest (e.g., "Europe"). This
parameter is ignored if level is set to 'cube' or 'world'. (Default: NULL)}
    \item{\code{ne_type}}{(Optional) The type of Natural Earth data to download:
'countries', 'map_units', 'sovereignty', or 'tiny_countries'. This parameter
is ignored if level is set to 'cube' or 'world'. (Default: "countries")}
    \item{\code{ne_scale}}{(Optional) The scale of Natural Earth data to download:
'small' - 110m, 'medium' - 50m, or 'large' - 10m. (Default: "medium")}
    \item{\code{output_crs}}{(Optional) The CRS you want for your calculated indicator.
(Leave blank to let the function choose a default based on grid reference
system.)}
    \item{\code{first_year}}{(Optional) Exclude data before this year. (Uses all data in
the cube by default.)}
    \item{\code{last_year}}{(Optional) Exclude data after this year. (Uses all data in
the cube by default.)}
    \item{\code{spherical_geometry}}{(Optional) If set to FALSE, will temporarily
disable spherical geometry while the function runs. Should only be used to
solve specific issues. (Default is TRUE).}
    \item{\code{make_valid}}{(Optional) Calls st_make_valid() from the sf package
after creating the grid. Increases processing time but may help if you are
getting polygon errors. (Default is FALSE).}
    \item{\code{shapefile_path}}{(optional) Path of an external shapefile to merge into
the workflow. For example, if you want to calculate your indicator
particular features such as protected areas or wetlands.}
    \item{\code{shapefile_crs}}{(Optional) CRS of a .wkt shapefile. If your shapefile
is .wkt and you do NOT use this parameter, the CRS will be assumed to be
EPSG:4326 and the coordinates will be read in as lat/long. If your shape is
NOT a .wkt the CRS will be determined automatically.}
    \item{\code{invert}}{(optional) Calculate an indicator over the inverse of the
shapefile (e.g. if you have a protected areas shapefile this would calculate
an indicator over all non protected areas within your cube). Default is
FALSE.}
    \item{\code{include_land}}{(Optional) Include occurrences which fall within the
land area. Default is TRUE. *Note that this purely a geographic filter, and
does not filter based on whether the occurrence is actually terrestrial.
Grid cells which fall partially on land and partially on ocean will be
included even if include_land is FALSE. To exclude terrestrial and/or
freshwater taxa, you must manually filter your data cube before calculating
your indicator.}
    \item{\code{include_ocean}}{(Optional) Include occurrences which fall outside the
land area. Default is TRUE. Set as "buffered_coast" to include a set buffer
size around the land area rather than the entire ocean area. *Note that this
is purely a geographic filter, and does not filter based on whether the
occurrence is actually marine. Grid cells which fall partially on land and
partially on ocean will be included even if include_ocean is FALSE. To
exclude marine taxa, you must manually filter your data cube before
calculating your indicator.}
    \item{\code{buffer_dist_km}}{(Optional) The distance to buffer around the land if
include_ocean is set to "buffered_coast". Default is 50 km.}
    \item{\code{force_grid}}{(Optional) Forces the calculation of a grid even if this
would not normally be part of the pipeline, e.g. for time series. This
setting is required for the calculation of rarity or Hill diversity, and is
forced on by indicators that require it. (Default: FALSE)}
  }}
}
\value{
An S3 object with the classes 'indicator_map' or 'indicator_ts' and
'total_occ' containing the calculated indicator values and metadata.
}
\description{
This function calculates the total number of species occurrence
records over a gridded map or as a time series (see 'Details' for more '
information).
}
\details{
The total number of occurrences is calculated by summing the occurrences of
all species observed for each cell or year. While not itself an indicator,
this variable provides an overview of the comprehensiveness and distribution
of data in the cube being analysed, and may be helpful, or even vital, for
interpreting the results of calculated indicators.
}
\section{Functions}{
\itemize{
\item \code{total_occ_map()}: 

\item \code{total_occ_ts()}: 

}}
\examples{
\donttest{
to_map <- total_occ_map(example_cube_1,
  level = "country",
  region = "Denmark"
)
plot(to_map)
}
\donttest{
to_ts <- total_occ_ts(example_cube_1, first_year = 1985)
plot(to_ts)
}
}
\seealso{
\code{\link[=compute_indicator_workflow]{compute_indicator_workflow()}}, \code{\link[=add_ci]{add_ci()}}
}
