% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{williams_evenness_map}
\alias{williams_evenness_map}
\title{Calculate Williams' Evenness Over Space}
\usage{
williams_evenness_map(
  data,
  cell_size = NULL,
  level = NULL,
  region = NULL,
  cube_crs = NULL,
  first_year = NULL,
  last_year = NULL,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{data}}{A data cube object (class 'processed_cube').}
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{cube_crs}}{The projection of the cube. (Default: "EPSG:3035")}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
  }}
}
\value{
An S3 object with the classes 'indicator_map' and 'williams_evenness' containing
the calculated indicator values and metadata.
}
\description{
This function calculates Williams' evenness over a gridded map.
}
\examples{
we_map <- williams_evenness_map(example_cube_1, level = "country", region = "Denmark")
plot(we_map)

}
\seealso{
compute_indicator_workflow
}
