% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{ab_rarity_ts}
\alias{ab_rarity_ts}
\title{Calculate Abundance-Based Rarity Over Time}
\usage{
ab_rarity_ts(data, ...)
}
\arguments{
\item{x}{A data cube object (class 'processed_cube').}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_ts' and 'ab_rarity' containing
the calculated indicator values and metadata.
}
\description{
This function calculates abundance-based rarity over time.
}
\examples{
abr_ts <- ab_rarity_ts(example_cube_1, first_year = 1985)
plot(abr_ts)

}
