% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3gbi-package.R
\docType{package}
\name{b3gbi-package}
\alias{b3gbi}
\alias{b3gbi-package}
\title{b3gbi: General Biodiversity Indicators for Biodiversity Data Cubes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate general biodiversity indicators from GBIF data cubes. Includes many common indicators such as species richness, evenness, and completeness, which can be calculated over time (trends) or space (maps). Now also supports retaining original cell IDs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-cubed-eu/b3gbi}
  \item \url{https://b-cubed-eu.github.io/b3gbi/}
  \item Report bugs at \url{https://github.com/b-cubed-eu/b3gbi/issues}
}

}
\author{
\strong{Maintainer}: Shawn Dove \email{Shawn.Dove@allzool.bio.uni-giessen.de} (\href{https://orcid.org/0000-0001-9465-5638}{ORCID})

Other contributors:
\itemize{
  \item Ward Langeraert (\href{https://orcid.org/0000-0002-5900-8109}{ORCID}) [contributor]
  \item Peter Desmet (\href{https://orcid.org/0000-0002-4442-8796}{ORCID}) [contributor]
  \item University of Giessen [copyright holder]
  \item European Union [funder]
}

}
\keyword{internal}
