% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3gbi-package.R
\docType{package}
\name{b3gbi-package}
\alias{b3gbi}
\alias{b3gbi-package}
\title{b3gbi: B-Cubed: General Biodiversity Indicators}
\description{
Calculate general biodiversity indicators from GBIF data cubes. Includes many common indicators such as species richness and evenness, which can be calculated over time (trends) or space (maps).
}
\author{
\strong{Maintainer}: Shawn Dove \email{s.dove@ucl.ac.uk}

}
\keyword{internal}
