% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci_methods.R
\name{calc_ci}
\alias{calc_ci}
\alias{calc_ci.obs_richness}
\alias{calc_ci.cum_richness}
\alias{calc_ci.total_occ}
\alias{calc_ci.occ_density}
\alias{calc_ci.newness}
\alias{calc_ci.williams_evenness}
\alias{calc_ci.pielou_evenness}
\alias{calc_ci.ab_rarity}
\alias{calc_ci.area_rarity}
\alias{calc_ci.spec_occ}
\alias{calc_ci.spec_range}
\alias{calc_ci.tax_distinct}
\alias{calc_ci.occ_turnover}
\title{Calculate Confidence Intervals for a Biodiversity Indicator}
\usage{
calc_ci(x, indicator, ...)

\method{calc_ci}{obs_richness}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{cum_richness}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{total_occ}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{occ_density}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{newness}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{williams_evenness}(x, ...)

\method{calc_ci}{pielou_evenness}(x, ...)

\method{calc_ci}{ab_rarity}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{area_rarity}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{spec_occ}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{spec_range}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

\method{calc_ci}{tax_distinct}(
  x,
  indicator,
  num_bootstrap = 1000,
  ci_type = ci_type,
  set_rows = 1,
  ...
)

\method{calc_ci}{occ_turnover}(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)
}
\arguments{
\item{x}{A data cube object}

\item{indicator}{The calculated indicator to generate confidence intervals for.
Only needed if you are using the function to generate bootstrap confidence
intervals (Default = NULL).}

\item{num_bootstrap}{Number of bootstraps to calculate for bootstrap confidence
intervals (Default = 1000).}

\item{ci_type}{Type of bootstrap confidence intervals to generate. Options are
'norm', 'basic', 'perc' for percent, or 'bca'.}
}
\description{
This function calculates bootstrap confidence intervals for a
biodiversity indicator. It is called automatically when calculating a biodiversity
indicator over time unless you choose 'none' for ci_type.
}
