% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci_methods.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate Confidence Intervals for a Biodiversity Indicator}
\usage{
calc_ci(x, indicator, ...)
}
\arguments{
\item{x}{A data cube object}

\item{indicator}{An indicator calculated over time, in the form of a data frame.
*Note: this should NOT be an 'indicator_ts' object as it is meant to be called by
the 'compute_indicator_workflow' function.}

\item{...}{Additional arguments passed to specific indicator calculation functions.}

\item{ci_type}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}

\item{num_bootstrap}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}

\item{set_rows}{Automatically select which taxonomic information to keep when
there are multiple options. Default value of 1 keeps the first option, which
is usually the best.}
}
\description{
This function calculates bootstrap confidence intervals for a
biodiversity indicator. It is called automatically when calculating a biodiversity
indicator over time unless you choose 'none' for ci_type.
}
