% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_map_methods.R, R/generic_functions.R
\name{calc_map.obs_richness}
\alias{calc_map.obs_richness}
\alias{calc_map.total_occ}
\alias{calc_map.newness}
\alias{calc_map.occ_density}
\alias{calc_map.williams_evenness}
\alias{calc_map.pielou_evenness}
\alias{calc_map.ab_rarity}
\alias{calc_map.area_rarity}
\alias{calc_map}
\title{Calculate Biodiversity Indicators Over Space}
\usage{
\method{calc_map}{obs_richness}(x, ...)

\method{calc_map}{total_occ}(x, ...)

\method{calc_map}{newness}(x, newness_min_year = NULL, ...)

\method{calc_map}{occ_density}(x, ...)

\method{calc_map}{williams_evenness}(x, ...)

\method{calc_map}{pielou_evenness}(x, ...)

\method{calc_map}{ab_rarity}(x, ...)

\method{calc_map}{area_rarity}(x, ...)

calc_map(x, ...)
}
\arguments{
\item{x}{A data cube object ('processed_cube').}

\item{...}{Additional arguments passed to specific indicator calculation functions.}

\item{type}{The indicator to calculate. Supported options include:
\itemize{
\item 'hill0', 'hill1', 'hill2': Hill numbers (order 0, 1, and 2).
\item 'obs_richness': Observed species richness.
\item 'total_occ': Total number of occurrences.
\item 'newness': Mean year of occurrence.
\item 'occ_density': Density of occurrences.
\item 'williams_evenness', 'pielou_evenness': Evenness measures.
\item 'ab_rarity', 'area_rarity':  Abundance-based and area-based rarity scores.
\item 'spec_occ': Species occurrences.
\item 'tax_distinct': Taxonomic distinctness.
}}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}
}
\value{
An S3 object of the class 'indicator_map' containing the calculated indicator values and metadata.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators on a spatial grid or as a time series. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_map'). Specific implementations
for different indicator types are provided using the appropriate wrappers.
}
\examples{
# Assuming 'my_data_cube' is a 'processed_cube' object
diversity_map <- obs_richness_map(my_data_cube, level = "continent", region = "Africa")

}
