% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_map_methods.R, R/generic_functions.R
\name{calc_map.obs_richness}
\alias{calc_map.obs_richness}
\alias{calc_map.total_occ}
\alias{calc_map.williams_evenness}
\alias{calc_map.pielou_evenness}
\alias{calc_map}
\title{Calculate a Biodiversity Indicator Map}
\usage{
\method{calc_map}{obs_richness}(x, ...)

\method{calc_map}{total_occ}(x, ...)

\method{calc_map}{williams_evenness}(x, ...)

\method{calc_map}{pielou_evenness}(x, ...)

calc_map(x)
}
\arguments{
\item{x}{An object containing the geospatial data and other information
required for the map calculation.}

\item{...}{Additional arguments (potentially used by specific methods).}
}
\value{
An object representing the calculated biodiversity indicator map.
The format of the output will depend on the specific S3 method used.
}
\description{
A generic function for calculating spatial biodiversity
indicator maps. Specific implementations for different indicator types
or mapping approaches should be provided using S3 methods.
}
