% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts.R
\name{calc_ts}
\alias{calc_ts}
\title{Calculate species richness trend}
\usage{
calc_ts(
  data,
  method = "observed",
  qval = 0,
  inext_sampsize = 150,
  coverage = 0.9
)
}
\arguments{
\item{data}{A tibble created from a GBIF cube using the process_cube function.}

\item{method}{A character vector.
"observed" calculates the observed species richness trend from the cube data,
"total_records" uses the total number of occurrences for each year as a
proxy for sampling effort.
"rarefaction" calculates rarefaction curves to approximate sampling effort,
"coverage" estimates relative richness trends by standardizing by coverage,
"evenness" calculates evenness}
}
\value{
A tibble
}
\description{
Calculate species richness trend
}
