% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_methods.R, R/generic_functions.R
\name{calc_ts.obs_richness}
\alias{calc_ts.obs_richness}
\alias{calc_ts.total_occ}
\alias{calc_ts.williams_evenness}
\alias{calc_ts.pielou_evenness}
\alias{calc_ts}
\title{Calculate a Biodiversity Indicator Time Series}
\usage{
\method{calc_ts}{obs_richness}(x, ...)

\method{calc_ts}{total_occ}(x, ...)

\method{calc_ts}{williams_evenness}(x, ...)

\method{calc_ts}{pielou_evenness}(x, ...)

calc_ts(x)
}
\arguments{
\item{x}{An object containing the necessary data and information for the
time series calculation.}

\item{...}{Additional arguments (potentially used by specific methods).}
}
\value{
An object representing the calculated biodiversity indicator time
series. The format of the output will depend on the specific S3 method used.
}
\description{
A generic function for calculating biodiversity indicator
time series data. Specific implementations for different indicator types
or calculation methods should be provided using S3 methods.
}
