% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_methods.R, R/generic_functions.R
\name{calc_ts.obs_richness}
\alias{calc_ts.obs_richness}
\alias{calc_ts.cum_richness}
\alias{calc_ts.total_occ}
\alias{calc_ts.occ_density}
\alias{calc_ts.newness}
\alias{calc_ts.williams_evenness}
\alias{calc_ts.pielou_evenness}
\alias{calc_ts.ab_rarity}
\alias{calc_ts.area_rarity}
\alias{calc_ts.spec_occ}
\alias{calc_ts.spec_range}
\alias{calc_ts.tax_distinct}
\alias{calc_ts.occ_turnover}
\alias{calc_ts}
\title{Calculate Biodiversity Indicators Over Time}
\usage{
\method{calc_ts}{obs_richness}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{cum_richness}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{total_occ}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{occ_density}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{newness}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{williams_evenness}(x, ...)

\method{calc_ts}{pielou_evenness}(x, ...)

\method{calc_ts}{ab_rarity}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{area_rarity}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{spec_occ}(
  x,
  indicator = indicator,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{spec_range}(
  x,
  indicator = indicator,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{tax_distinct}(
  x,
  set_rows = 1,
  indicator = indicator,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

\method{calc_ts}{occ_turnover}(
  x,
  indicator = NULL,
  bootstrap = FALSE,
  num_bootstrap = 1000,
  ci_type = ci_type,
  ...
)

calc_ts(x, ...)
}
\arguments{
\item{x}{A data cube object ('processed_cube').}

\item{indicator}{The calculated indicator to generate confidence intervals for.
Only needed if you are using the function to generate bootstrap confidence
intervals (Default = NULL).}

\item{bootstrap}{Set this to true to generate bootstrap confidence intervals
(Default = FALSE).}

\item{num_bootstrap}{Number of bootstraps to calculate for bootstrap confidence
intervals (Default = 1000).}

\item{ci_type}{Type of bootstrap confidence intervals to generate. Options are
'norm', 'basic', 'perc' for percent, or 'bca'.}

\item{...}{Additional arguments passed to specific indicator calculation functions.}

\item{set_rows}{Automatically select which taxonomic information to keep when
there are multiple options. Default value of 1 keeps the first option, which
is usually the best.}
}
\value{
An S3 object of the class 'indicator_ts' containing the calculated indicator values and metadata.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators over time. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_ts').
Specific implementations for different indicator types are provided using the
appropriate wrappers.
}
\examples{
occurrence_density_trend <- occ_density_ts(example_cube_1)
plot(occurrence_density_trend, min_year=1980)

}
