% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_methods.R, R/generic_functions.R
\name{calc_ts.obs_richness}
\alias{calc_ts.obs_richness}
\alias{calc_ts.cum_richness}
\alias{calc_ts.total_occ}
\alias{calc_ts.occ_density}
\alias{calc_ts.newness}
\alias{calc_ts.williams_evenness}
\alias{calc_ts.pielou_evenness}
\alias{calc_ts.ab_rarity}
\alias{calc_ts.area_rarity}
\alias{calc_ts}
\title{Calculate Biodiversity Indicators Over Time}
\usage{
\method{calc_ts}{obs_richness}(x, ...)

\method{calc_ts}{cum_richness}(x, ...)

\method{calc_ts}{total_occ}(x, ...)

\method{calc_ts}{occ_density}(x, ...)

\method{calc_ts}{newness}(x, ...)

\method{calc_ts}{williams_evenness}(x, ...)

\method{calc_ts}{pielou_evenness}(x, ...)

\method{calc_ts}{ab_rarity}(x, ...)

\method{calc_ts}{area_rarity}(x, ...)

calc_ts(x, ...)
}
\arguments{
\item{x}{A data cube object ('processed_cube').}

\item{...}{Additional arguments passed to specific indicator calculation functions.}

\item{type}{The indicator to calculate. Supported options include:
\itemize{
\item 'obs_richness': Observed species richness.
\item 'cum_richness': Cumulative species richness.
\item 'total_occ': Total number of occurrences.
\item 'newness': Mean year of occurrence.
\item 'occ_density': Density of occurrences.
\item 'williams_evenness', 'pielou_evenness': Evenness measures.
\item 'ab_rarity', 'area_rarity':  Abundance-based and area-based rarity scores.
}}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}
}
\value{
An S3 object of the class 'indicator_ts' containing the calculated indicator values and metadata.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators over time. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_ts').
Specific implementations for different indicator types are provided using the
appropriate wrappers.
}
\examples{
occurrence_density_trend <- occ_density_ts(example_cube_1)
plot(occurrence_density_trend, min_year=1980)

}
