% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_methods.R, R/generic_functions.R
\name{calc_ts.hill0}
\alias{calc_ts.hill0}
\alias{calc_ts.hill1}
\alias{calc_ts.hill_core}
\alias{calc_ts.obs_richness}
\alias{calc_ts.cum_richness}
\alias{calc_ts.total_occ}
\alias{calc_ts.occ_density}
\alias{calc_ts.newness}
\alias{calc_ts.williams_evenness}
\alias{calc_ts.pielou_evenness}
\alias{calc_ts.evenness_core}
\alias{calc_ts.ab_rarity}
\alias{calc_ts.area_rarity}
\alias{calc_ts.spec_occ}
\alias{calc_ts.spec_range}
\alias{calc_ts.tax_distinct}
\alias{calc_ts.occ_turnover}
\alias{calc_ts}
\title{Calculate Biodiversity Indicators Over Time}
\usage{
\method{calc_ts}{hill0}(x, ...)

\method{calc_ts}{hill1}(x, ...)

\method{calc_ts}{hill_core}(x, type = c("hill0", "hill1", "hill2"), ...)

\method{calc_ts}{obs_richness}(x, ...)

\method{calc_ts}{cum_richness}(x, ...)

\method{calc_ts}{total_occ}(x, ...)

\method{calc_ts}{occ_density}(x, ...)

\method{calc_ts}{newness}(x, ...)

\method{calc_ts}{williams_evenness}(x, ...)

\method{calc_ts}{pielou_evenness}(x, ...)

\method{calc_ts}{evenness_core}(x, type, ...)

\method{calc_ts}{ab_rarity}(x, ...)

\method{calc_ts}{area_rarity}(x, ...)

\method{calc_ts}{spec_occ}(x, ...)

\method{calc_ts}{spec_range}(x, ...)

\method{calc_ts}{tax_distinct}(x, set_rows = 1, ...)

\method{calc_ts}{occ_turnover}(x, ...)

calc_ts(x, ...)
}
\arguments{
\item{x}{A data cube object ('processed_cube').}

\item{...}{Additional arguments passed to specific indicator calculation functions.}

\item{type}{Choose which Hill number, or q, to calculate. Choose 'hill0' (q = 0)
for estimated species richness, 'hill1' for Hill-Shannon diversity, or 'hill2'
for Hill-Simpson diversity.}

\item{set_rows}{Automatically select which taxonomic information to keep when
there are multiple options. Default value of 1 keeps the first option, which
is usually the best.}
}
\value{
An S3 object of the class 'indicator_ts' containing the calculated indicator values and metadata.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators over time. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_ts').
Specific implementations for different indicator types are provided using the
appropriate wrappers.
}
\examples{
total_occurrences_trend <- total_occ_ts(example_cube_1)
plot(total_occurrences_trend, min_year=1980)

}
