% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{completeness_map}
\alias{completeness_map}
\alias{completeness_ts}
\title{Calculate Completeness (Sample Coverage) Over Space or Time}
\usage{
completeness_map(
  data,
  cutoff_length = 5,
  data_type = c("incidence", "abundance"),
  assume_freq = FALSE,
  ...
)

completeness_ts(data, cutoff_length = 5, gridded_average = FALSE, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{cutoff_length}{(Optional) The minimum number of species or observations
required for a grid cell or time point to be included. Default is 5.}

\item{data_type}{The type of data: "incidence" or "abundance". Default is "incidence".}

\item{assume_freq}{(Optional) Whether to assume frequency data if using
incidence. Default is FALSE.}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{cell_size}}{(Optional) Length of grid cell sides, in km or degrees.
If set to "grid" (default), this will use the existing grid size of your
cube. If set to "auto", this will be automatically determined according to
the geographical level selected. This is 100 km or 1 degree for 'continent'
or 'world', 10 km or (for a degree-based CRS) the native resolution of the
cube for 'country', 'sovereignty' or 'geounit'. If level is set to 'cube',
cell size will be the native resolution of the cube for a degree-based CRS,
or for a km-based CRS, the cell size will be determined by the area of the
cube: 100 km for cubes larger than 1 million sq km, 10 km for cubes between
10 thousand and 1 million sq km, 1 km for cubes between 100 and 10 thousand
sq km, and 0.1 km for cubes smaller than 100 sq km. Alternatively, the user
can manually select the grid cell size (in km or degrees). Note that the
cell size must be a whole number multiple of the cube's resolution.}
    \item{\code{level}}{(Optional) Spatial level: 'cube', 'continent', 'country',
'world', 'sovereignty', or 'geounit'. (Default: 'cube')}
    \item{\code{region}}{(Optional) The region of interest (e.g., "Europe"). This
parameter is ignored if level is set to 'cube' or 'world'. (Default: NULL)}
    \item{\code{ne_type}}{(Optional) The type of Natural Earth data to download:
'countries', 'map_units', 'sovereignty', or 'tiny_countries'. This parameter
is ignored if level is set to 'cube' or 'world'. (Default: "countries")}
    \item{\code{ne_scale}}{(Optional) The scale of Natural Earth data to download:
'small' - 110m, 'medium' - 50m, or 'large' - 10m. (Default: "medium")}
    \item{\code{output_crs}}{(Optional) The CRS you want for your calculated indicator.
(Leave blank to let the function choose a default based on grid reference
system.)}
    \item{\code{first_year}}{(Optional) Exclude data before this year. (Uses all data in
the cube by default.)}
    \item{\code{last_year}}{(Optional) Exclude data after this year. (Uses all data in
the cube by default.)}
    \item{\code{spherical_geometry}}{(Optional) If set to FALSE, will temporarily
disable spherical geometry while the function runs. Should only be used to
solve specific issues. (Default is TRUE).}
    \item{\code{make_valid}}{(Optional) Calls st_make_valid() from the sf package
after creating the grid. Increases processing time but may help if you are
getting polygon errors. (Default is FALSE).}
    \item{\code{shapefile_path}}{(optional) Path of an external shapefile to merge into
the workflow. For example, if you want to calculate your indicator
particular features such as protected areas or wetlands.}
    \item{\code{shapefile_crs}}{(Optional) CRS of a .wkt shapefile. If your shapefile
is .wkt and you do NOT use this parameter, the CRS will be assumed to be
EPSG:4326 and the coordinates will be read in as lat/long. If your shape is
NOT a .wkt the CRS will be determined automatically.}
    \item{\code{invert}}{(optional) Calculate an indicator over the inverse of the
shapefile (e.g. if you have a protected areas shapefile this would calculate
an indicator over all non protected areas within your cube). Default is
FALSE.}
    \item{\code{include_land}}{(Optional) Include occurrences which fall within the
land area. Default is TRUE. *Note that this purely a geographic filter, and
does not filter based on whether the occurrence is actually terrestrial.
Grid cells which fall partially on land and partially on ocean will be
included even if include_land is FALSE. To exclude terrestrial and/or
freshwater taxa, you must manually filter your data cube before calculating
your indicator.}
    \item{\code{include_ocean}}{(Optional) Include occurrences which fall outside the
land area. Default is TRUE. Set as "buffered_coast" to include a set buffer
size around the land area rather than the entire ocean area. *Note that this
is purely a geographic filter, and does not filter based on whether the
occurrence is actually marine. Grid cells which fall partially on land and
partially on ocean will be included even if include_ocean is FALSE. To
exclude marine taxa, you must manually filter your data cube before
calculating your indicator.}
    \item{\code{buffer_dist_km}}{(Optional) The distance to buffer around the land if
include_ocean is set to "buffered_coast". Default is 50 km.}
    \item{\code{force_grid}}{(Optional) Forces the calculation of a grid even if this
would not normally be part of the pipeline, e.g. for time series. This
setting is required for the calculation of rarity or Hill diversity, and is
forced on by indicators that require it. (Default: FALSE)}
  }}

\item{gridded_average}{(Optional) For time series, calculate completeness
for each grid cell and average the results, rather than calculating for
the entire area at once. Default is FALSE.}
}
\value{
An S3 object with the classes 'indicator_map' or 'indicator_ts' and
'completeness' containing the calculated indicator values and metadata.
}
\description{
This function calculates the completeness of biodiversity data
over a gridded map or as a time series using the Sample Coverage metric
from the iNEXT package.
}
\details{
\subsection{Completeness (Sample Coverage)}{

Completeness is measured as \strong{Sample Coverage}, a concept developed by Turing and Good (1953) and further popularized in ecology by Chao and Jost (2012). Sample coverage estimates the proportion of the total individuals in an ecological community that belong to the species detected in a sample.

A coverage value of 1.0 indicates that no new species are expected to be uncovered through further sampling, meaning the sample is perfectly complete. A value of 0.8, for example, suggests that 20\% of the individuals in the community belong to species that have not yet been detected in the sample.

Calculating sample coverage is a prerequisite for reliable comparisons of biodiversity across different areas or time periods, as it provides a standardized measure of sample completeness that is independent of sample size alone (Chao et al., 2014).

In this package, completeness is calculated using the 'iNEXT' package based on the observed data in each grid cell or time point.
}
}
\section{Functions}{
\itemize{
\item \code{completeness_map()}: 

\item \code{completeness_ts()}: 

}}
\examples{
\donttest{
comp_map <- completeness_map(example_cube_1)
plot(comp_map)
}
\donttest{
comp_ts <- completeness_ts(example_cube_1, first_year = 1985)
plot(comp_ts)
}
}
