% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{compute_indicator_workflow}
\alias{compute_indicator_workflow}
\title{Calculate Biodiversity Indicators Over Space or Time}
\usage{
compute_indicator_workflow(
  data,
  type,
  dim_type = c("map", "ts"),
  cell_size = NULL,
  level = c("continent", "country", "world"),
  region = "Europe",
  cube_crs = NULL,
  first_year = NULL,
  last_year = NULL,
  ...
)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{type}{The indicator to calculate. Supported options include:
\itemize{
\item 'obs_richness': Observed species richness.
\item 'total_occ': Total number of occurrences.
\item 'newness': Mean year of occurrence.
\item 'density': Density of occurrences.
\item 'williams_evenness', 'pielou_evenness': Evenness measures.
\item 'ab_rarity', 'area_rarity':  Abundance-based and area-based rarity scores.
}}

\item{dim_type}{Dimension to calculate indicator over (time: 'ts', or space: 'map')}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}

\item{...}{Additional arguments passed to specific indicator calculation functions.}
}
\value{
An S3 object containing the calculated indicator values and metadata.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators on a spatial grid or as a time series. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_map' or 'indicator_ts').
}
\examples{
diversity_map <- compute_indicator_workflow(example_cube_2,
                                            type = "obs_richness",
                                            dim_type = "map",
                                            level = "continent",
                                            region = "Europe")
diversity_map

}
