% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_indicator_workflow.R
\name{compute_indicator_workflow}
\alias{compute_indicator_workflow}
\title{Calculate Biodiversity Indicators Over Space or Time}
\usage{
compute_indicator_workflow(
  x,
  type,
  dim_type = c("map", "ts"),
  cell_size = NULL,
  level = c("continent", "country", "world"),
  region = "Europe",
  ...
)
}
\arguments{
\item{x}{A data cube object ('processed_cube', 'processed_cube_dsinfo', or  'virtual_cube').}

\item{type}{The indicator to calculate. Supported options include:
\itemize{
\item 'hill0', 'hill1', 'hill2': Hill numbers (order 0, 1, and 2).
\item 'obs_richness': Observed species richness.
\item 'total_occ': Total number of occurrences.
\item 'newness': Mean year of occurrence.
\item 'density': Density of occurrences.
\item 'e9_evenness', 'pielou_evenness': Evenness measures.
\item 'ab_rarity', 'area_rarity':  Abundance-based and area-based rarity scores.
\item 'spec_occ': Species occurrences.
\item 'tax_distinct': Taxonomic distinctness.
}}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{...}{Additional arguments passed to specific indicator calculation functions.}
}
\value{
An S3 object of the appropriate class containing the calculated indicator values and metadata:
\itemize{
\item 'indicator_map' for real-world observational data calculated over a grid (map)
\item 'indicator_ts' for real-world observational data calculated over time (time series)
\item 'virtual_indicator_map' for virtual species data calculated over a grid (map).
}
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators on a spatial grid. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_map' or 'virtual_indicator_map').
}
\examples{
# Assuming 'my_data_cube' is a 'processed_cube' or 'virtual_cube' object
diversity_map <- calculate_indicator(my_data_cube, type = "obs_richness", level = "continent", region = "Africa")

}
