% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{compute_indicator_workflow}
\alias{compute_indicator_workflow}
\title{Calculate Biodiversity Indicators Over Space or Time}
\usage{
compute_indicator_workflow(
  data,
  type,
  dim_type = c("map", "ts"),
  ci_type = c("norm", "basic", "perc", "bca", "none"),
  cell_size = NULL,
  level = c("cube", "continent", "country", "world", "sovereignty", "geounit"),
  region = "Europe",
  ne_type = c("countries", "map_units", "sovereignty", "tiny_countries"),
  ne_scale = c("medium", "small", "large"),
  output_crs = NULL,
  first_year = NULL,
  last_year = NULL,
  spherical_geometry = TRUE,
  make_valid = FALSE,
  num_bootstrap = 100,
  shapefile_path = NULL,
  shapefile_crs = NULL,
  invert = FALSE,
  include_ocean = TRUE,
  buffer_dist_km = 50,
  force_grid = FALSE,
  ...
)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{type}{The indicator to calculate. Supported options include:
\itemize{
\item 'obs_richness': Observed species richness.
\item 'total_occ': Total number of occurrences.
\item 'newness': Mean year of occurrence.
\item 'occ_density': Density of occurrences.
\item 'williams_evenness', 'pielou_evenness': Evenness measures.
\item 'ab_rarity', 'area_rarity':  Abundance-based and area-based rarity scores.
\item 'cum_richness': Cumulative species richness.
\item 'occ_turnover': Occupancy turnover.
\item 'spec_range': Species range size.
\item 'spec_occ': Species occurrences.
\item 'tax_distinct': Taxonomic distinctness.
\item 'hill0': Species richness (estimated by coverage-based rarefaction).
\item 'hill1': Hill-Shannon diversity (estimated by coverage-based rarefaction).
\item 'hill2': Hill-Simpson diversity (estimated by coverage-based rarefaction).
}}

\item{dim_type}{Dimension to calculate indicator over (time: 'ts', or space: 'map')}

\item{ci_type}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'cube', 'continent', 'country', 'world', 'sovereignty',
or 'geounit'. (Default: 'cube')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{ne_type}{The type of Natural Earth data to download: 'countries', 'map_units',
'sovereignty', or 'tiny_countries'. (Default: "countries")}

\item{ne_scale}{The scale of Natural Earth data to download: 'small' - 110m,
'medium' - 50m, or 'large' - 10m. (Default: "medium")}

\item{output_crs}{The CRS you want for your calculated indicator. (Leave blank
to let the function choose a default based on grid reference system)}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}

\item{spherical_geometry}{If set to FALSE, will temporarily disable spherical geometry
while the function runs. Should only be used to solve specific issues. (Default is TRUE)}

\item{make_valid}{Calls st_make_valid() from the sf package. Increases processing
time but may help if you are getting polygon errors. (Default is FALSE).}

\item{num_bootstrap}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}

\item{shapefile_path}{Path of an external shapefile to merge into the workflow. For example,
if you want to calculate your indicator particular features such as protected areas or wetlands.}

\item{shapefile_crs}{CRS of a .wkt shapefile. If your shapefile is .wkt and you do
NOT use this parameter, the CRS will be assumed to be EPSG:4326 and the
coordinates will be read in as lat/long. If your shape is NOT a .wkt the CRS
will be determined automatically.}

\item{invert}{Calculate an indicator over the inverse of the shapefile (e.g.
if you have a protected areas shapefile this would calculate an indicator over
all non protected areas)}

\item{include_ocean}{Include occurrences which fall outside the land area.
Default is TRUE. Set as "buffered_coast" to include a set buffer size around
the land area rather than the entire ocean area.}

\item{buffer_dist_km}{The distance to buffer around the land if include_ocean
is set to "buffered_coast".}

\item{force_grid}{Forces the calculation of a grid even if this would not
normally be part of the pipeline, e.g. for time series. This setting is
required for the calculation of rarity, and is turned on by the ab_rarity_ts
and area_rarity_ts wrappers. (Default: FALSE)}

\item{...}{Additional arguments passed to specific indicator calculation functions.}
}
\value{
An S3 object containing the calculated indicator values and metadata.
}
\description{
This function provides a flexible framework for calculating various biodiversity
indicators on a spatial grid or as a time series. It prepares the data, creates a grid, calculates indicators,
and formats the output into an appropriate S3 object ('indicator_map' or 'indicator_ts').
}
\examples{
diversity_map <- compute_indicator_workflow(example_cube_1,
                                            type = "obs_richness",
                                            dim_type = "map",
                                            level = "country",
                                            region = "Denmark")
diversity_map

}
