% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sf_from_utm.R
\name{create_sf_from_utm}
\alias{create_sf_from_utm}
\title{Create an sf object from UTM coordinates, handling multiple zones correctly.}
\usage{
create_sf_from_utm(df, output_crs = NULL)
}
\arguments{
\item{df}{A data frame with columns: xcoord, ycoord, and utmzone.}

\item{output_crs}{(Optional) The EPSG code or CRS string for the desired
output CRS. If NULL, the CRS of the first UTM zone will be used.}
}
\value{
An sf object with the geometry correctly defined for each UTM zone.
}
\description{
This function takes a data frame with UTM coordinates (xcoord, ycoord) and a
utmzone column, and creates an sf object with the correct CRS for each zone.
}
