% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{cum_richness_ts}
\alias{cum_richness_ts}
\title{Calculate Cumulative Species Richness}
\usage{
cum_richness_ts(data, ...)
}
\arguments{
\item{x}{A data cube object (class 'processed_cube').}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_ts' and 'cum_richness' containing
the calculated indicator values and metadata.
}
\description{
This function calculates cumulative species richness as a time series.
It prepares the data, creates a grid, calculates the indicator value for each grid cell,
and formats the output into an S3 object.
}
\examples{
cr_ts <- cum_richness_ts(example_cube_1)
plot(cr_ts)

}
