% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{cum_richness_ts}
\alias{cum_richness_ts}
\title{Calculate Cumulative Species Richness}
\usage{
cum_richness_ts(data, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{ci_type}}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'cube', 'continent', 'country', 'world', 'sovereignty',
or 'geounit'. (Default: 'cube')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{ne_type}}{The type of Natural Earth data to download: 'countries', 'map_units',
'sovereignty', or 'tiny_countries'. (Default: "countries")}
    \item{\code{ne_scale}}{The scale of Natural Earth data to download: 'small' - 110m,
'medium' - 50m, or 'large' - 10m. (Default: "medium")}
    \item{\code{output_crs}}{The CRS you want for your calculated indicator. (Leave blank
to let the function choose a default based on grid reference system)}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
    \item{\code{spherical_geometry}}{If set to FALSE, will temporarily disable spherical geometry
while the function runs. Should only be used to solve specific issues. (Default is TRUE)}
    \item{\code{make_valid}}{Calls st_make_valid() from the sf package. Increases processing
time but may help if you are getting polygon errors. (Default is FALSE).}
    \item{\code{num_bootstrap}}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}
    \item{\code{crs_unit_convert}}{Force a particular output CRS even when it has
different units than the input CRS. (Default: FALSE)}
    \item{\code{shapefile_path}}{Path of an external shapefile to merge into the workflow. For example,
if you want to calculate your indicator particular features such as protected areas or wetlands.}
    \item{\code{invert}}{Calculate an indicator over the inverse of the shapefile (e.g.
if you have a protected areas shapefile this would calculate an indicator over
all non protected areas)}
  }}
}
\value{
An S3 object with the classes 'indicator_ts' and 'cum_richness' containing
the calculated indicator values and metadata.
}
\description{
This function calculates cumulative species richness as a time
series (see 'Details' for more information).
}
\details{
\if{html}{\out{
<h3>Species richness</h3>
Species richness is the total number of species present in a
sample (Magurran, 1988). It is a fundamental and commonly used
measure of biodiversity, providing a simple and intuitive
overview of the status of biodiversity. However, richness is not
well suited to measuring biodiversity change over time, as it
only decreases when local extinctions occur and thus lags behind
abundance for negative trends. While it may act as a leading
indicator of alien species invasions, it will not indicate
establishment because it ignores abundance. Nor will it
necessarily indicate changes in local species composition, which
can occur without any change in richness. Although richness is
conceptually simple, it can be measured in different ways.
}}
\if{html}{\out{
<h3>Cumulative richness</h3>
Cumulative richness is calculated by adding the newly observed unique species
each year to a cumulative sum. This indicator provides an estimation of
whether and how many new species are still being discovered in a region.
While an influx of alien species could cause an increase in cumulative
richness, a fast-rising trend as shown in Fig. 2 is likely an indication
that the dataset is not comprehensive and therefore observed richness will
provide an underestimate of species richness.
}}
}
\examples{
\dontrun{
cr_ts <- cum_richness_ts(example_cube_1, first_year = 1985)
plot(cr_ts)
}
}
\seealso{
compute_indicator_workflow
}
