% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cube_1}
\alias{example_cube_1}
\title{Cube of GBIF Mammal Occurrences in Denmark}
\format{
A 'processed_cube' object: a list of metadata plus, in its \code{data}
element, a tibble with 31,632 rows and 15 variables
\describe{
\item{year}{year occurrence was recorded}
\item{cellCode}{code containing the cell resolution and coordinates
on the Extended Quarter Degree (eqdgc) grid}
\item{kingdomKey}{kingdom key associated with the species on GBIF}
\item{kingdom}{kingdom name}
\item{familyKey}{family key associated with the species on GBIF}
\item{family}{family name}
\item{taxonKey}{taxonomic key associated with the species on GBIF}
\item{scientificName}{scientific species name}
\item{obs}{number of occurrences (records)}
\item{minCoordinateUncertaintyInMeters}{minimum coordinate uncertainty in
meters}
\item{minTemporalUncertainty}{minimum temporal uncertainty in seconds}
\item{familyCount}{number of occurrences of the family the species
belongs to}
\item{xcoord}{longitude of the cell centre (degrees)}
\item{ycoord}{latitude of the cell centre (degrees)}
\item{resolution}{grid cell size (e.g. "0.25degrees")}
}
}
\source{
\doi{10.15468/dl.5mb887}
}
\usage{
example_cube_1
}
\description{
Small example cube containing the mammal occurrences in Denmark available on
GBIF as of 16.03.2024.
}
\examples{
\donttest{
denmark_mammals_or_map <- obs_richness_map(example_cube_1,
                                           level = "country",
                                           region = "Denmark")
plot(denmark_mammals_or_map,
     title = "Mammals in Denmark (1862-2024): Observed Species Richness")
}
}
\keyword{datasets}
