% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_indicator_map1}
\alias{example_indicator_map1}
\title{Map of Pielou's Evenness for Mammals in Denmark}
\format{
An 'indicator_map' object containing a tibble with 576 rows and 4 variables, as well as metadata
\describe{
\item{cellid}{ID number identifying a particular grid cell}
\item{area_km2}{area of the grid cell in square kms}
\item{diversity_val}{calculated evenness value for the grid cell}
\item{geometry}{grid cell polygon (this column is not visible by default when you print the object)}
}
}
\source{
\url{https://doi.org/10.15468/dl.qk4f2z}
}
\usage{
example_indicator_map1
}
\description{
Example indicator containing a map of pielou's evenness for mammal occurrences in Denmark (occurrences from GBIF: 16.03.2024).
}
\keyword{datasets}
