% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_indicator_map1}
\alias{example_indicator_map1}
\title{Map of Observed Species Richness for Mammals in Denmark}
\format{
An 'indicator_map' object containing a tibble with 800 rows and 5
variables, as well as metadata
\describe{
\item{cellid}{id of a map cell the indicator was calculated for}
\item{area}{area of the map cell in square kilometers}
\item{cellCode}{code containing the cell coordinates in Extended Quarter
Degree (eqdgc) grid format}
\item{diversity_val}{calculated richness value for the cell}
\item{geometry}{geometry of the map cell}
}
}
\source{
\url{https://doi.org/10.15468/dl.5mb887}
}
\usage{
example_indicator_map1
}
\description{
Example indicator containing a map of observed species richness for mammal
occurrences in Denmark (occurrences from GBIF: 16.03.2024).
}
\keyword{datasets}
