% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_indicator_ts1}
\alias{example_indicator_ts1}
\title{Time Series of Observed Species Richness for Mammals in Denmark}
\format{
An 'indicator_ts' object containing a tibble with 55 rows and 2
variables, as well as metadata
\describe{
\item{year}{a year the indicator was calculated for}
\item{diversity_val}{calculated richness value for the year}
}
}
\source{
\doi{10.15468/dl.5mb887}
}
\usage{
example_indicator_ts1
}
\description{
Example indicator containing a time series of observed species richness for
mammal occurrences in Denmark, 1970-2024 (occurrences from GBIF: 16.03.2024).
}
\keyword{datasets}
