% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_indicator_ts2}
\alias{example_indicator_ts2}
\title{Time Series of Cumulative Species Richness for Bryophytes in South Africa}
\format{
An 'indicator_ts' object containing a tibble with 110 rows and 2
variables, as well as metadata
\describe{
\item{year}{a year the indicator was calculated for}
\item{diversity_val}{calculated richness value for the year}
}
}
\source{
\url{https://doi.org/10.15468/dl.yfzgja}
}
\usage{
example_indicator_ts2
}
\description{
Example indicator containing a time series of cumulative species richness for
bryophyte occurrences in South Africa, 1875-2024 (occurrences from GBIF:
16.03.2024).
}
\keyword{datasets}
