% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_species.R
\name{list_species}
\alias{list_species}
\title{Extract species names from a 'processed_cube', 'indicator_map', or 'indicator_ts' object.}
\usage{
list_species(object)
}
\arguments{
\item{object}{A biodiversity data object containing species names, either as
a separate vector called species_names or as a column called scientificName.}
}
\value{
Either a character vector of species names (if directly available)
or a data frame with columns:
\itemize{
\item \code{taxonKey}: A unique identifier for each species.
\item \code{scientificName}:  The scientific name of each species.
}
}
\description{
Retrieves a list of species names from a biodiversity data object.
}
\examples{
# Assuming you have a biodiversity data cube named 'mammals_cube'
species_list <- list_species(mammals_cube)
}
