% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrs_to_latlong_bbox.R
\name{mgrs_to_latlong_bbox}
\alias{mgrs_to_latlong_bbox}
\title{Create a Single Bounding Box from MGRS Data in a Projected CRS}
\usage{
mgrs_to_latlong_bbox(df)
}
\arguments{
\item{df}{A data frame with at least three columns: \code{cellCode} (containing
MGRS strings), and \code{xcoord} and \code{ycoord} for easting and northing.}
}
\value{
An sf bounding box (\code{sf::st_bbox}) in a single projected CRS (e.g., Albers).
}
\description{
Converts MGRS coordinates to a single sf bounding box in a
suitable projected CRS, handling data that spans multiple UTM zones.
}
