% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{newness_map}
\alias{newness_map}
\alias{newness_ts}
\title{Calculate Mean Year of Occurrence Over Time or Space}
\usage{
newness_map(data, ...)

newness_ts(data, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{ci_type}}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'cube', 'continent', 'country', 'world', 'sovereignty',
or 'geounit'. (Default: 'cube')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{ne_type}}{The type of Natural Earth data to download: 'countries', 'map_units',
'sovereignty', or 'tiny_countries'. (Default: "countries")}
    \item{\code{ne_scale}}{The scale of Natural Earth data to download: 'small' - 110m,
'medium' - 50m, or 'large' - 10m. (Default: "medium")}
    \item{\code{output_crs}}{The CRS you want for your calculated indicator. (Leave blank
to let the function choose a default based on grid reference system)}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
    \item{\code{spherical_geometry}}{If set to FALSE, will temporarily disable spherical geometry
while the function runs. Should only be used to solve specific issues. (Default is TRUE)}
    \item{\code{make_valid}}{Calls st_make_valid() from the sf package. Increases processing
time but may help if you are getting polygon errors. (Default is FALSE).}
    \item{\code{num_bootstrap}}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}
    \item{\code{crs_unit_convert}}{Force a particular output CRS even when it has
different units than the input CRS. (Default: FALSE)}
    \item{\code{shapefile_path}}{Path of an external shapefile to merge into the workflow. For example,
if you want to calculate your indicator particular features such as protected areas or wetlands.}
    \item{\code{invert}}{Calculate an indicator over the inverse of the shapefile (e.g.
if you have a protected areas shapefile this would calculate an indicator over
all non protected areas)}
    \item{\code{include_water}}{Include rnaturalearth oceans, rivers, and lakes layers.
Default is TRUE. Set as "buffered_coast" to include a set buffer size around
the land area.}
    \item{\code{buffer_dist_km}}{The distance to buffer around the land if include_water
is set to "buffered_coast".}
    \item{\code{layers}}{Names of additional rnaturalearth layers you want to add to your map.
E.g. c("reefs", "playas", "admin_0_boundary_lines_land")}
  }}
}
\value{
An S3 object with the classes 'indicator_map' or 'indicator_ts' and
'newness' containing the calculated indicator values and metadata.
}
\description{
This function estimates the relative newness of records in a
data cube by calculating the mean year of occurrence over a gridded map or as
a time series (see 'Details' for more information).
}
\details{
The mean year of occurrence is calculated per cell, giving an
indication of how recent the data is for each cell. A recent
mean year is not necessarily an indication of quality, as some
countries or regions have been conducting comprehensive
biodiversity monitoring for many years and will therefore
reflect an older mean year of occurrence, while others may show
a recent mean year due to e.g., the sudden availability of large
amounts of citizen science data.
}
\section{Functions}{
\itemize{
\item \code{newness_map()}: 

\item \code{newness_ts()}: 

}}
\examples{
\dontrun{
n_map <- newness_map(example_cube_1, level = "country", region = "Denmark")
plot(n_map)
}
\dontrun{
n_ts <- newness_ts(example_cube_1, first_year = 1985)
plot(n_ts)
}
}
\seealso{
compute_indicator_workflow
}
