% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{newness_map}
\alias{newness_map}
\title{Calculate Mean Year of Occurrence Over Space}
\usage{
newness_map(data, ...)
}
\arguments{
\item{x}{A data cube object (class 'processed_cube').}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_map' and 'newness' containing
the calculated indicator values and metadata.
}
\description{
This function estimates the relative newness of records in a data cube
by calculating the mean year of occurrence over a gridded map.
}
\examples{
n_map <- newness_map(example_cube_1, level = "country", region = "Denmark")
plot(n_map)

}
