% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{occ_turnover_ts}
\alias{occ_turnover_ts}
\title{Calculate Occupancy Turnover}
\usage{
occ_turnover_ts(data, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{ci_type}}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'cube', 'continent', 'country', 'world', 'sovereignty',
or 'geounit'. (Default: 'cube')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{ne_type}}{The type of Natural Earth data to download: 'countries', 'map_units',
'sovereignty', or 'tiny_countries'. (Default: "countries")}
    \item{\code{ne_scale}}{The scale of Natural Earth data to download: 'small' - 110m,
'medium' - 50m, or 'large' - 10m. (Default: "medium")}
    \item{\code{output_crs}}{The CRS you want for your calculated indicator. (Leave blank
to let the function choose a default based on grid reference system)}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
    \item{\code{spherical_geometry}}{If set to FALSE, will temporarily disable spherical geometry
while the function runs. Should only be used to solve specific issues. (Default is TRUE)}
    \item{\code{make_valid}}{Calls st_make_valid() from the sf package. Increases processing
time but may help if you are getting polygon errors. (Default is FALSE).}
    \item{\code{num_bootstrap}}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}
    \item{\code{crs_unit_convert}}{Force a particular output CRS even when it has
different units than the input CRS. (Default: FALSE)}
    \item{\code{shapefile_path}}{Path of an external shapefile to merge into the workflow. For example,
if you want to calculate your indicator particular features such as protected areas or wetlands.}
    \item{\code{invert}}{Calculate an indicator over the inverse of the shapefile (e.g.
if you have a protected areas shapefile this would calculate an indicator over
all non protected areas)}
  }}
}
\value{
An S3 object with the classes 'indicator_ts' and 'occ_turnover'
containing the calculated indicator values and metadata.
}
\description{
This function calculates occupancy turnover as a time series
(see 'Details' for more information).
}
\details{
Occupancy turnover measures the change in species composition over time,
reflecting the rate at which species appear or disappear from a given area.
It provides insights into the dynamic nature of ecological communities,
highlighting shifts in species distributions and potential environmental
changes. High turnover rates may indicate rapid community restructuring,
potentially driven by factors such as habitat alteration, climate change,
or invasive species. Analyzing occupancy turnover can be crucial for
understanding ecosystem stability, identifying areas of conservation concern,
and assessing the effectiveness of management strategies.

Occupancy turnover can be calculated in different ways, but here we use the
Jaccard dissimilarity index (Jaccard, 1901) to measure the similarity between
two sets of species occurrences. The Jaccard index is calculated as:

\deqn{
  J = (b + c) / (a + b + c)
}{
  J = (b + c) / (a + b + c)
}

where a is the number of species present in both time periods, b is the
number of species present only in the first time period, and c is the number
of species present only in the second time period. The index ranges
from 0 (no turnover) to 1 (complete turnover).
}
\examples{
ot_ts <- occ_turnover_ts(example_cube_1, first_year = 1985)
plot(ot_ts)

}
\references{
Jaccard, P. (1901). Étude de la distribution florale dans une portion
des Alpes et du Jura. \emph{Bulletin de la Société Vaudoise des}
\emph{Science Naturelles}, \emph{37}(142), 547-579.
}
\seealso{
compute_indicator_workflow
}
