% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{pielou_evenness_map}
\alias{pielou_evenness_map}
\title{Calculate Pielou's Evenness Over Space}
\usage{
pielou_evenness_map(data, ...)
}
\arguments{
\item{x}{A data cube object (class 'processed_cube').}

\item{cell_size}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}

\item{level}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}

\item{region}{The region of interest (e.g., "Europe"). (Default: "Europe")}

\item{cube_crs}{The projection of the cube. (Default: "EPSG:3035")}

\item{first_year}{Exclude data before this year. (Uses all data in the cube by default.)}

\item{last_year}{Exclude data after this year. (Uses all data in the cube by default.)}
}
\value{
An S3 object with the classes 'indicator_map' and 'pielou_evenness' containing
the calculated indicator values and metadata.
}
\description{
This function calculates Pielou's evenness over a gridded
map. It prepares the data, creates a grid, calculates the indicator value for
each grid cell, and formats the output into an S3 object.
}
\examples{
pe_map <- pielou_evenness_map(example_cube_1)
plot(pe_map)

}
