% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot_map}
\alias{plot_map}
\title{Plot Biodiversity Indicator Map}
\usage{
plot_map(
  x,
  title = "auto",
  auto_title = NULL,
  leg_label_default = NULL,
  xlims = NULL,
  ylims = NULL,
  trans = NULL,
  breaks = NULL,
  labels = NULL,
  Europe_crop = TRUE,
  surround = TRUE,
  panel_bg = NULL,
  legend_title = NULL,
  legend_limits = NULL,
  legend_title_wrap_length = 10,
  title_wrap_length = 60
)
}
\arguments{
\item{x}{An 'indicator_map' object containing indicator values associated with
map grid cells.}

\item{title}{Plot title. Replace "auto" with your own title if you want a
custom title or if calling the function manually.}

\item{auto_title}{Text for automatic title generation, provided by an
appropriate S3 method (if calling the function manually, leave as NULL).}

\item{leg_label_default}{Default label for the legend, provided by an
appropriate S3 method (if calling the function manually, leave as NULL).}

\item{xlims}{(Optional) Custom x-axis limits.}

\item{ylims}{(Optional) Custom y-axis limits.}

\item{trans}{(Optional) Scale transformation for the fill gradient
(e.g., 'log').}

\item{breaks}{(Optional) Break points for the legend scale.}

\item{labels}{(Optional) Labels for legend scale break points.}

\item{Europe_crop}{If TRUE, crops maps of Europe to exclude far-lying islands.}

\item{surround}{If TRUE, includes surrounding countries in gray when plotting
at the country level.}

\item{panel_bg}{(Optional) Background color for the map panel.}

\item{legend_title}{(Optional) Title for the plot legend.}

\item{legend_limits}{(Optional) Limits for the legend scale.}

\item{wrap_length}{Maximum title length before wrapping to a new line.}
}
\value{
A ggplot object representing the biodiversity indicator map.
}
\description{
Creates a map visualization of a calculated biodiversity indicator,
providing customization options.
}
\examples{
# Assuming you have an 'indicator_map' object named 'richness_map':
plot_map(x = richness_map, title = "Map of Species Richness")

}
