% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_map}
\alias{plot_map}
\title{Plot Biodiversity Indicator Map}
\usage{
plot_map(
  x,
  title = "auto",
  auto_title = NULL,
  leg_label_default = NULL,
  xlims = NULL,
  ylims = NULL,
  trans = NULL,
  bcpower = NULL,
  breaks = NULL,
  labels = NULL,
  crop_to_grid = FALSE,
  panel_bg = NULL,
  land_fill_colour = NULL,
  legend_title = NULL,
  legend_limits = NULL,
  legend_title_wrap_length = 10,
  title_wrap_length = 60,
  visible_gridlines = TRUE,
  layers = NULL,
  scale = c("medium", "small", "large")
)
}
\arguments{
\item{x}{An 'indicator_map' object containing indicator values associated
with map grid cells.}

\item{title}{(Optional) Plot title. Replace "auto" with your own title if you
want a custom title or if calling the function manually.}

\item{auto_title}{(Optional) Text for automatic title generation, provided by
an appropriate S3 method (if calling the function manually, leave as NULL).}

\item{leg_label_default}{(Optional) Default label for the legend, provided by
an appropriate S3 method (if calling the function manually, leave as NULL).}

\item{xlims}{(Optional) Custom x-axis limits.}

\item{ylims}{(Optional) Custom y-axis limits.}

\item{trans}{(Optional) Scale transformation for the fill gradient
(e.g., 'log').}

\item{bcpower}{(Optional) Power parameter for the Box-Cox, modulus, or
Yeo-Johnson transformations.}

\item{breaks}{(Optional) Break points for the legend scale.}

\item{labels}{(Optional) Labels for legend scale break points.}

\item{crop_to_grid}{(Optional) If TRUE, the grid will determine the edges of
the map. Default is FALSE.}

\item{panel_bg}{(Optional) Background colour for the map panel.}

\item{land_fill_colour}{(Optional) Colour for the land area outside of the
grid (if surround = TRUE). Default is "grey85".}

\item{legend_title}{(Optional) Title for the plot legend.}

\item{legend_limits}{(Optional) Limits for the legend scale.}

\item{legend_title_wrap_length}{(Optional) Maximum legend title length before
wrapping to a new line.}

\item{title_wrap_length}{(Optional) Maximum title length before wrapping to a
new line.}

\item{visible_gridlines}{(Optional) Show gridlines between cells. Default is
TRUE.}

\item{layers}{(Optional) Additional rnaturalearth layers to plot, e.g.
c("reefs", "playas").}

\item{scale}{(Optional) Scale of Natural Earth data ("small", "medium", or
"large"). Default is 'medium'.}
}
\value{
A ggplot object representing the biodiversity indicator map.
Can be customized using ggplot2 functions.
}
\description{
Creates a map visualization of a calculated biodiversity
indicator, providing customization options.
}
\examples{
evenness_map <- pielou_evenness_map(example_cube_1,
                                    level = "country",
                                    region = "Denmark")
plot_map(x = evenness_map,
         title = "Map of Species Evenness in Denmark",
         legend_title = "Evenness")

}
