% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_species_map}
\alias{plot_species_map}
\title{Plot Occurrence Map or Range Map of Individual Species}
\usage{
plot_species_map(
  x,
  species = NULL,
  leg_label_default = NULL,
  auto_title = NULL,
  suppress_legend = FALSE,
  title = "auto",
  xlims = NULL,
  ylims = NULL,
  trans = NULL,
  breaks = NULL,
  labels = NULL,
  min_year = NULL,
  max_year = NULL,
  smoothed_trend = TRUE,
  linecolour = NULL,
  trendlinecolour = NULL,
  envelopecolour = NULL,
  Europe_crop = TRUE,
  surround = TRUE,
  single_plot = TRUE,
  panel_bg = NULL,
  x_label = NULL,
  y_label = NULL,
  x_breaks = 10,
  y_breaks = 6,
  suppress_y = FALSE,
  gridoff = FALSE,
  legend_title = NULL,
  legend_limits = NULL,
  legend_title_wrap_length = 10,
  title_wrap_length = 60,
  ...
)
}
\arguments{
\item{x}{An 'indicator_map' object containing indicator values associated with
map grid cells.}

\item{leg_label_default}{Default label for the legend, provided by an
appropriate S3 method (if calling the function manually, leave as NULL).}

\item{auto_title}{Text for automatic title generation, provided by an
appropriate S3 method (if calling the function manually, leave as NULL).}

\item{suppress_legend}{Do not show legend. This should be set to true when
plotting species ranges, as all cell values are 1.}

\item{title}{Plot title. Replace "auto" with your own title if you want a
custom title or if calling the function manually.}

\item{xlims}{(Optional) Custom x-axis limits.}

\item{ylims}{(Optional) Custom y-axis limits.}

\item{trans}{(Optional) Scale transformation for the fill gradient
(e.g., 'log').}

\item{breaks}{(Optional) Break points for the legend scale.}

\item{labels}{(Optional) Labels for legend scale break points.}

\item{Europe_crop}{If TRUE, crops maps of Europe to exclude far-lying islands.}

\item{surround}{If TRUE, includes surrounding countries in gray when plotting
at the country level.}

\item{panel_bg}{(Optional) Background color for the map panel.}

\item{legend_title}{(Optional) Title for the plot legend.}

\item{legend_limits}{(Optional) Limits for the legend scale.}

\item{legend_title_wrap_length}{Maximum legend title length before wrapping to a new line.}

\item{title_wrap_length}{Maximum title length before wrapping to a new line.}
}
\value{
A ggplot object representing the map of species range or occurrences.
Can be customized using ggplot2 functions.
}
\description{
Creates map visualizations of species ranges or species occurrences,
providing customization options.
}
\examples{
spec_occ_map <- pielou_evenness_map(example_cube_1,
                                    level = "country",
                                    region = "Denmark")
plot_map(x = spec_occ_map, c(2440728, 4265185))

}
