% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_species_ts}
\alias{plot_species_ts}
\title{Plot Occurrence Trends or Range Size Trends for Individual Species}
\usage{
plot_species_ts(
  x,
  species = NULL,
  y_label_default = NULL,
  auto_title = NULL,
  title = "auto",
  xlims = NULL,
  ylims = NULL,
  trans = NULL,
  breaks = NULL,
  labels = NULL,
  min_year = NULL,
  max_year = NULL,
  smoothed_trend = TRUE,
  linecolour = NULL,
  trendlinecolour = NULL,
  envelopecolour = NULL,
  Europe_crop = TRUE,
  surround = TRUE,
  single_plot = TRUE,
  panel_bg = NULL,
  x_label = NULL,
  y_label = NULL,
  x_breaks = 10,
  y_breaks = 6,
  suppress_y = FALSE,
  gridoff = FALSE,
  legend_title = NULL,
  legend_limits = NULL,
  legend_title_wrap_length = 10,
  title_wrap_length = 60,
  ...
)
}
\arguments{
\item{x}{An 'indicator_ts' object containing time series of indicator values
matched to species names and/or taxon keys.}

\item{y_label_default}{Default label for the y-axis, provided by an appropriate
S3 method (if calling the function manually, leave as NULL).}

\item{auto_title}{Text for automatic title generation, provided by an
appropriate S3 method (if calling the function manually, leave as NULL).}

\item{title}{Plot title. Replace "auto" with your own title if you want a
custom title or if calling the function manually.}

\item{min_year}{(Optional)  Earliest year to include in the plot.}

\item{max_year}{(Optional)  Latest year to include in the plot.}

\item{smoothed_trend}{If TRUE, plot a smoothed trendline.}

\item{linecolour}{(Optional) Colour for the indicator line.
Default is darkorange.}

\item{trendlinecolour}{(Optional) Colour for the smoothed trendline.
Default is blue.}

\item{envelopecolour}{(Optional) Colour for the uncertainty envelope.
Default is lightsteelblue.}

\item{x_label}{Label for the x-axis.}

\item{y_label}{Label for the y-axis.}

\item{x_breaks}{Integer giving desired number of breaks for x axis.
(May not return exactly the number requested.)}

\item{y_breaks}{Integer giving desired number of breaks for y axis.
(May not return exactly the number requested.)}

\item{suppress_y}{If TRUE, suppresses y-axis labels.}

\item{gridoff}{If TRUE, hides gridlines.}

\item{wrap_length}{Maximum title length before wrapping to a new line.}
}
\value{
A ggplot object representing species range or occurrence time series plot(s).
Can be customized using ggplot2 functions.
}
\description{
Creates time series plots of species occurrences or species range
sizes, with an optional smoothed trendline, and visualizes uncertainty.
}
\examples{
spec_occ_ts_mammals_denmark <- spec_occ_ts(example_cube_1,
                                        level = "country",
                                        region = "Denmark")
# default colours:
plot_species_ts(spec_occ_ts_mammals_denmark, c(2440728, 4265185))

# custom colours:
plot_species_ts(spec_occ_ts_mammals_denmark, c(2440728, 4265185),
        linecolour = "thistle",
        trendlinecolour = "forestgreen",
        envelopecolour = "lightgreen")
}
