% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot Biodiversity Indicator Trend}
\usage{
plot_ts(
  x,
  title = "auto",
  auto_title = NULL,
  y_label_default = NULL,
  suppress_y = FALSE,
  smoothed_trend = TRUE,
  linecolour = NULL,
  trendlinecolour = NULL,
  envelopecolour = NULL,
  gridoff = FALSE,
  x_label = NULL,
  y_label = NULL,
  min_year = NULL,
  max_year = NULL,
  x_breaks = 10,
  y_breaks = 6,
  wrap_length = 60
)
}
\arguments{
\item{x}{An 'indicator_ts' object containing a time series of indicator values.}

\item{title}{Plot title. Replace "auto" with your own title if you want a
custom title or if calling the function manually.}

\item{auto_title}{Text for automatic title generation, provided by an
appropriate S3 method (if calling the function manually, leave as NULL).}

\item{y_label_default}{Default label for the y-axis, provided by an appropriate
S3 method (if calling the function manually, leave as NULL).}

\item{suppress_y}{If TRUE, suppresses y-axis labels.}

\item{smoothed_trend}{If TRUE, plot a smoothed trendline.}

\item{linecolour}{(Optional) Colour for the indicator line.
Default is darkorange.}

\item{trendlinecolour}{(Optional) Colour for the smoothed trendline.
Default is blue.}

\item{envelopecolour}{(Optional) Colour for the uncertainty envelope.
Default is lightsteelblue.}

\item{gridoff}{If TRUE, hides gridlines.}

\item{x_label}{Label for the x-axis.}

\item{y_label}{Label for the y-axis.}

\item{min_year}{(Optional)  Earliest year to include in the plot.}

\item{max_year}{(Optional)  Latest year to include in the plot.}

\item{x_breaks}{Integer giving desired number of breaks for x axis.
(May not return exactly the number requested.)}

\item{y_breaks}{Integer giving desired number of breaks for y axis.
(May not return exactly the number requested.)}

\item{wrap_length}{Maximum title length before wrapping to a new line.}
}
\value{
A ggplot object representing the biodiversity indicator time series plot.
Can be customized using ggplot2 functions.
}
\description{
Creates a time series plot of a calculated biodiversity
indicator, with an optional smoothed trendline, and visualizes uncertainty.
}
\examples{
# default colours:
plot_ts(example_indicator_ts2, y_label = "Species Richness", title = "Cumulative Species Richness: Insects in Europe")

# custom colours:
plot_ts(example_indicator_ts2,
        y_label = "Species Richness",
        title = "Cumulative Species Richness: Insects in Europe",
        linecolour = "thistle",
        trendlinecolour = "forestgreen",
        envelopecolour = "lightgreen")
}
