% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.available_indicators}
\alias{print.available_indicators}
\title{Print Available Indicators}
\usage{
\method{print}{available_indicators}(x, n = 30, ...)
}
\arguments{
\item{x}{Object of class available_indicators}

\item{n}{(Optional) Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provide a summary of indicators registered in the package.
}
