% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.indicator_map}
\alias{print.indicator_map}
\title{Print an Indicator Map Object}
\usage{
\method{print}{indicator_map}(x, n = 10, include_na = FALSE, ...)
}
\arguments{
\item{x}{An indicator_map object.}

\item{n}{Integer specifying the number of rows of data to display.}

\item{include_na}{Logical. If TRUE, includes rows with NA diversity values
in the printed output. Default is FALSE.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a summary representation of an indicator_map object,
designed for user-friendly display in the console.
}
