% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.processed_cube}
\alias{print.processed_cube}
\title{Print a Processed Data Cube Object}
\usage{
\method{print}{processed_cube}(x, n = 10, ...)
}
\arguments{
\item{x}{A processed_cube object.}

\item{n}{Integer specifying the number of rows of cube data to display.}

\item{...}{Additional arguments.}
}
\description{
Provides a summary representation of a processed_cube object,
designed for user-friendly display in the console.
}
