% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cube.R
\name{process_cube}
\alias{process_cube}
\title{Process a GBIF data cube}
\usage{
process_cube(
  cube_name,
  tax_info,
  datasets_info = NA,
  first_year = NA,
  final_year = NA
)
}
\arguments{
\item{cube_name}{The location and name of a data cube file to open.}

\item{tax_info}{The location and name of an associated taxonomic info file.}

\item{datasets_info}{The location and name of an associated dataset info file.}

\item{first_year}{The first year of occurrences (if different from cube).}

\item{final_year}{The final year of occurrences (if different from cube).}
}
\value{
A tibble.
}
\description{
Process a GBIF data cube
}
