% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cube_functions.R
\name{process_cube}
\alias{process_cube}
\title{Process GBIF Data Cubes}
\usage{
process_cube(
  cube_name,
  tax_info = NULL,
  datasets_info = NULL,
  first_year = 1600,
  last_year = NULL
)
}
\arguments{
\item{cube_name}{The location and name of a data cube file
(e.g., 'inst/extdata/europe_species_cube.csv').}

\item{tax_info}{The location and name of an associated taxonomic info file
(e.g.,  'inst/extdata/europe_species_info.csv').}

\item{datasets_info}{The location and name of an associated dataset info file
(e.g., 'inst/extdata/europe_species_datasets.csv').}

\item{first_year}{(Optional) The first year of occurrences to include. If not
specified, uses a default of 1600 to prevent false records (e.g. with year = 0).}

\item{last_year}{(Optional) The final year of occurrences to include. If not
specified, uses the latest year present in the cube.}
}
\value{
A tibble containing the processed GBIF occurrence data.
}
\description{
Processes a GBIF data cube and associated taxonomic information
file. All files must reside in the same directory and share a base file
name (e.g., 'my_mammals_cube.csv', 'my_mammals_info.csv').
}
\examples{
\dontrun{
cube_name <- system.file("extdata", "europe_species_cube.csv", package = "b3gbi")
tax_info <- system.file("extdata", "europe_species_info.csv", package = "b3gbi")
europe_example_cube <- process_cube(cube_name, tax_info)
europe_example_cube
}
}
