% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cube_functions.R
\name{process_cube}
\alias{process_cube}
\alias{process_cube_new}
\title{Process GBIF Data Cubes}
\usage{
process_cube(
  cube_name,
  tax_info = NULL,
  datasets_info = NULL,
  first_year = 1600,
  last_year = NULL,
  ...
)

process_cube_new(
  cube_name,
  grid_type = c("automatic", "eea", "mgrs", "eqdgc"),
  first_year = NULL,
  last_year = NULL,
  force_gridcode = FALSE,
  cols_year = NULL,
  cols_yearmonth = NULL,
  cols_cellcode = NULL,
  cols_occurrences = NULL,
  cols_scientificname = NULL,
  cols_mincoordinateuncertaintyinmeters = NULL,
  cols_mintemporaluncertainty = NULL,
  cols_kingdom = NULL,
  cols_family = NULL,
  cols_species = NULL,
  cols_kingdomkey = NULL,
  cols_familykey = NULL,
  cols_specieskey = NULL,
  cols_familycount = NULL
)
}
\arguments{
\item{cube_name}{The location and name of a data cube file
(e.g., 'inst/extdata/europe_species_cube.csv').}

\item{tax_info}{The location and name of an associated taxonomic info file
(e.g.,  'inst/extdata/europe_species_info.csv').}

\item{datasets_info}{The location and name of an associated dataset info file
(e.g., 'inst/extdata/europe_species_datasets.csv').}

\item{first_year}{(Optional) The first year of occurrences to include. If not
specified, uses a default of 1600 to prevent false records (e.g. with year = 0).}

\item{last_year}{(Optional) The final year of occurrences to include. If not
specified, uses the latest year present in the cube.}

\item{...}{Passes arguments to process_cube_new if you are using a cube generated
with the new GBIF API.}

\item{grid_type}{Specify which grid reference system your cube uses. By default
the function will attempt to determine this automatically and return an error if it fails.}

\item{cols_year}{The name of the column containing the year of occurrence (if
something other than 'year'). This column is required unless you have a yearMonth
column.}

\item{cols_yearmonth}{The name of the column containing the year and month of
occurrence (if present and if other than 'yearMonth'). Use this if only if you
do not have a year column. The b3gbi package does not use month data, so the
function will convert your yearMonth column to a year column.}

\item{cols_cellcode}{The name of the column containing the grid reference codes
(if other than 'cellCode'). This column is required.}

\item{cols_occurrences}{The name of the column containing the number of occurrence
(if other than 'occurrences'). This column is required.}

\item{cols_scientificname}{The name of the column containing the scientific name
of the species (if other than 'scientificName'). Note that it is not necessary
to have both a species column and a scientificName column. One or the other is
sufficient.}

\item{cols_mincoordinateuncertaintyinmeters}{The name of the column containing
the minimum coordinate uncertainty of the occurrences (if other than
'minCoordinateUncertaintyinMeters').}

\item{cols_mintemporaluncertainty}{The name of the column containing the minimum
temporal uncertainty of the occurrences (if other than 'minTemporalUncertainty').}

\item{cols_kingdom}{The name of the column containing the kingdom the occurring
species belongs to (if other than 'kingdom'). This column is optional.}

\item{cols_family}{The name of the column containing the family the occurring
species belongs to (if other than 'family'). This column is optional.}

\item{cols_species}{The name of the column containing the name of the occurring
species (if other than 'species'). Note that it is not necessary to have both a
species column and a scientificName column. One or the other is sufficient.}

\item{cols_kingdomkey}{The name of the column containing the kingdom key of the
occurring species (if other than 'kingdomKey'). This column is optinal.}

\item{cols_familykey}{The name of the column containing the family key of the
occurring species (if other than 'familykey'). This column is optional.}

\item{cols_specieskey}{The name of the column containing the species key of the
occurring species (if other than 'speciesKey'). This column is required, but
note that if you have a 'taxonKey' column you can provide it as the speciesKey.}

\item{cols_familycount}{The name of the column containing the occurrence count
by family. This column is optional.}
}
\value{
A tibble containing the processed GBIF occurrence data.
}
\description{
Processes a GBIF data cube and (if applicable) an associated taxonomic
information file. If your cube includes a taxonomic info file it is likely a
previous generation cube. The taxonomic info file must reside in the same directory
as your cube and share a base file name (e.g., 'cubes/my_mammals_cube.csv', 'cubes/my_mammals_info.csv').
If your cube does NOT include a taxonomic info file then it is likely a current
generation cube and will be automatically passed to the process_cube_new function.
As cube generation is more flexible for these cubes, you may need to provide
column names for some columns if they differ from the default names expected
by the function.
}
\examples{
\dontrun{
cube_name <- system.file("extdata", "europe_species_cube.csv", package = "b3gbi")
tax_info <- system.file("extdata", "europe_species_info.csv", package = "b3gbi")
europe_example_cube <- process_cube(cube_name, tax_info)
europe_example_cube
}
}
