% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{tax_distinct_map}
\alias{tax_distinct_map}
\title{Calculate Taxonomic Distinctness Over Space}
\usage{
tax_distinct_map(data, rows = 1, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{rows}{Choose which row to select if there are multiple matches when retrieving taxonomic information from GBIF. (Default is 1. Use NA for interactive mode.)}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{ci_type}}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{output_crs}}{The CRS you want for your calculated indicator. (Leave blank
to let the function choose a default based on grid reference system)}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
    \item{\code{spherical_geometry}}{If set to FALSE, will temporarily disable spherical geometry
while the function runs. Should only be used to solve specific issues. (Default is TRUE)}
    \item{\code{make_valid}}{Calls st_make_valid() from the sf package. Increases processing
time but may help if you are getting polygon errors. (Default is FALSE).}
    \item{\code{num_bootstrap}}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}
  }}
}
\value{
An S3 object with the classes 'indicator_map' and 'tax_distinct' containing
the calculated indicator values and metadata.
}
\description{
This function calculates the taxonomic distinctness index over a gridded map.
}
\examples{
\dontrun{
td_map <- tax_distinct_map(example_cube_1, level = "country", region = "Denmark")
plot(td_map)
}

}
\seealso{
compute_indicator_workflow
}
