% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_wrappers.R
\name{tax_distinct_map}
\alias{tax_distinct_map}
\alias{tax_distinct_ts}
\title{Calculate Taxonomic Distinctness Over Space or Time}
\usage{
tax_distinct_map(data, rows = 1, ...)

tax_distinct_ts(data, rows = 1, ...)
}
\arguments{
\item{data}{A data cube object (class 'processed_cube').}

\item{rows}{Choose which row to select if there are multiple matches when
retrieving taxonomic information from GBIF. (Default is 1. Use NA for
interactive mode.)}

\item{...}{
  Arguments passed on to \code{\link[=compute_indicator_workflow]{compute_indicator_workflow}}
  \describe{
    \item{\code{ci_type}}{Type of bootstrap confidence intervals to calculate. (Default: "norm".
Select "none" to avoid calculating bootstrap CIs.)}
    \item{\code{cell_size}}{Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)}
    \item{\code{level}}{Spatial level: 'cube', 'continent', 'country', 'world', 'sovereignty',
or 'geounit'. (Default: 'cube')}
    \item{\code{region}}{The region of interest (e.g., "Europe"). (Default: "Europe")}
    \item{\code{ne_type}}{The type of Natural Earth data to download: 'countries', 'map_units',
'sovereignty', or 'tiny_countries'. (Default: "countries")}
    \item{\code{ne_scale}}{The scale of Natural Earth data to download: 'small' - 110m,
'medium' - 50m, or 'large' - 10m. (Default: "medium")}
    \item{\code{output_crs}}{The CRS you want for your calculated indicator. (Leave blank
to let the function choose a default based on grid reference system)}
    \item{\code{first_year}}{Exclude data before this year. (Uses all data in the cube by default.)}
    \item{\code{last_year}}{Exclude data after this year. (Uses all data in the cube by default.)}
    \item{\code{spherical_geometry}}{If set to FALSE, will temporarily disable spherical geometry
while the function runs. Should only be used to solve specific issues. (Default is TRUE)}
    \item{\code{make_valid}}{Calls st_make_valid() from the sf package. Increases processing
time but may help if you are getting polygon errors. (Default is FALSE).}
    \item{\code{num_bootstrap}}{Set the number of bootstraps to calculate for generating
confidence intervals. (Default: 1000)}
    \item{\code{shapefile_path}}{Path of an external shapefile to merge into the workflow. For example,
if you want to calculate your indicator particular features such as protected areas or wetlands.}
    \item{\code{shapefile_crs}}{CRS of a .wkt shapefile. If your shapefile is .wkt and you do
NOT use this parameter, the CRS will be assumed to be EPSG:4326 and the
coordinates will be read in as lat/long. If your shape is NOT a .wkt the CRS
will be determined automatically.}
    \item{\code{invert}}{Calculate an indicator over the inverse of the shapefile (e.g.
if you have a protected areas shapefile this would calculate an indicator over
all non protected areas)}
    \item{\code{include_ocean}}{Include occurrences which fall outside the land area.
Default is TRUE. Set as "buffered_coast" to include a set buffer size around
the land area rather than the entire ocean area.}
    \item{\code{buffer_dist_km}}{The distance to buffer around the land if include_ocean
is set to "buffered_coast".}
  }}
}
\value{
An S3 object with the classes 'indicator_map' or 'indicator_ts' and
'tax_distinct' containing the calculated indicator values and metadata.
}
\description{
This function calculates the taxonomic distinctness index (TDI)
over a gridded map or as a time series (see 'Details' for more information).
}
\details{
Taxonomic distinctness is an essential biodiversity variable (EBV)
that measures the taxonomic relatedness between species, providing a measure
of biodiversity that accounts for evolutionary relationships. A distance
matrix based on pairwise taxonomic relationships is calculated for each cell
using the taxize package (Chamberlain & Szöcs, 2013; Chamberlain et al.,
2020), then taxonomic distinctness is calculated as the Taxonomic
Distinctness Index (TDI; Clarke & Warwick, 1999):
\deqn{
 \frac{\sum\sum_{i<j} \frac{|R_i - R_j|}{L}}{\frac{S(S-1)}{2}}
}{
 (&sum;&sum; from i<j of (|R_i-R_j| / L) / (S * (S - 1) / 2)
}
where S is the number of species, Ri and Rj are the taxonomic ranks
of species i and j (from the GBIF Taxonomic Backbone), and L is the
maximum number of taxonomic ranks.
The TDI ranges from 0 to 1, with higher values indicating greater
taxonomic distinctness.
}
\section{Functions}{
\itemize{
\item \code{tax_distinct_map()}: 

\item \code{tax_distinct_ts()}: 

}}
\examples{
\dontrun{
td_map <- tax_distinct_map(example_cube_1, level = "country", region = "Denmark")
plot(td_map)
}

\dontrun{
td_ts <- tax_distinct_ts(example_cube_1, level = "country", region = "Denmark")
plot(td_ts)
}
}
\references{
Chamberlain, S. A., & Szöcs, E. (2013). taxize: taxonomic search and
retrieval in R. \emph{F1000Research}, 2.

Chamberlain, S., Szoecs, E., Foster, Z., Boettiger, C., Ram, K., Bartomeus,
I., Baumgartner, J., O'Donnell, J., Oksanen, J., Tzovaras, B. G., Marchand,
P., Tran, V., Salmon, M., Li, G., & Grenié, M. (2020). taxize: Taxonomic
Information from Around the Web. R package version 0.9.98.
https://github.com/ropensci/taxize.

Clarke, K. R., & Warwick, R. M. (1999). The taxonomic distinctness measure
of biodiversity: weighting of step lengths between hierarchical levels.
Marine Ecology Progress Series, 184, 21-29.
}
\seealso{
compute_indicator_workflow
}
